feat(FR-2862): add preset mode for deployment revision creation (#7350)
Resolves #7347 (FR-2862)

## Test

```
10.122.10.215
```

`e2e-cpu-http-server` preset makes a helathy route.

## Summary

Implements **M2** and **M3** of FR-2862 ("Deployment Revision Preset Mode UX").

### M2 — VFolder Deploy: preset selector

- `VFolderDeployModal` now requires a deployment-revision **preset**; resource group + replicas/openToPublic are still configurable but image/runtime/cluster/resources/env are sourced from the selected preset.
- Submits via `deployVfolderV2` (single-shot deployment + initial revision) and polls the new endpoint until it appears in the v1 store.
- Empty-preset state: shows a banner with a link that opens the deployment preset creation modal directly.
- Entry points wired: `VFolderNodes`, `VFolderNodesV2`, `ModelCardDrawer`, `ModelCardDeployModal`.

### M3 — Add Revision: Preset / Custom modes

- `DeploymentAddRevisionModal` now opens in **Preset mode** by default; an antd `Segmented` switches to **Custom mode** (the legacy advanced form).
- Mode preference persists per-user via `useBAISettingUserState('deploymentRevisionCreationMode')`.
- Preset → Custom carries the user's preset selection forward by reading an `@inline` Relay fragment with `readInlineData` and prefilling the custom form (cluster, image, runtime, resources, environ, replicas, public flag).
- Removed the previous auto-prefill from the current revision. The Custom body now surfaces an `Alert` action ("Load current revision") that the user must click explicitly. The Alert is gated on the deployment having a current revision.

### Structural refactor

- **Split** the previous monolithic `DeploymentAddRevisionFormContents.tsx` into two co-located bodies:
    - `DeploymentAddRevisionCustomContent.tsx` — Custom-mode form + `addModelRevision` mutation.
    - `DeploymentAddRevisionPresetContent.tsx` — Preset-mode form + `deployVfolderV2` mutation.
Each body owns its own fragment, mutation, and form values type. The modal wrapper now only hosts the segmented control, the shared `useLazyLoadQuery`, and the Preset → Custom transfer bridge.
- **Removed** the now-unused legacy `DeploymentLauncherPage` / `DeploymentLauncherPageContent` (≈2,000 lines), whose responsibilities are subsumed by the modal-based flows above.
- **Validation guard**: explicit `form.validateFields().then(...)` immediately before each create/deploy mutation (`commitAdd` in Custom, `commitDeploy` in Preset) so the rule "validate before any deploy/commit mutation" holds at the mutation call site, not just at the `onFinish` boundary. The modal's `handleOk` also validates before `form.submit()`.

## Test plan

- [ ] VFolder → Deploy: empty-presets banner opens the create-preset modal.
- [ ] VFolder → Deploy: select a preset, submit; endpoint detail page opens after the v1 store catches up.
- [ ] Deployment detail → Add revision: opens in Preset mode by default.
- [ ] Switch Preset → Custom: carry-over preset values appear in the custom form.
- [ ] Toggle Preset ⇄ Custom multiple times; persisted mode survives reload.
- [ ] Custom mode: clicking "Load current revision" prefills the form; modal opened with no current revision hides the Alert.
- [ ] Preset / Custom: submitting with an invalid field scrolls to the first error and does not fire the mutation.
