+++ react/src/components/VFolderNodesV2.tsx
@@ -14,6 +14,7 @@ import { useSuspenseTanQuery, useTanQuery } from '../hooks/reactQueryAlias';
 import { useSetBAINotification } from '../hooks/useBAINotification';
 import { isDeletedCategory } from '../pages/VFolderNodeListPage';
 import DeleteForeverVFolderModalV2 from './DeleteForeverVFolderModalV2';
+import DeploymentSettingModal from './DeploymentSettingModal';
 import { useFolderExplorerOpener } from './FolderExplorerOpener';
 import InviteFolderSettingModal from './InviteFolderSettingModal';
 import QuotaPerStorageVolumePanelCard, {
@@ -29,6 +30,7 @@ import {
   QuestionCircleOutlined,
   UserOutlined,
 } from '@ant-design/icons';
+import { useToggle } from 'ahooks';
 import { App, Modal, Skeleton, theme, Tooltip, Typography } from 'antd';
 import {
   filterOutNullAndUndefined,
@@ -439,6 +441,11 @@ const VFolderNodesV2: React.FC<VFolderNodesV2Props> = ({
   const [deployFallbackVfolderId, setDeployFallbackVfolderId] = useState<
     string | null
   >(null);
+  // FR-2862 — when the user hits the empty-preset state in
+  // VFolderDeployModal, escalate to the deployment shell creation modal
+  // (`DeploymentSettingModal`), same as the `/deployments` page entry.
+  const [isCreateDeploymentOpen, { toggle: toggleCreateDeployment }] =
+    useToggle(false);
   const [isHostQuotaModalOpen, setIsHostQuotaModalOpen] = useState(false);
 
   // `vfolderStatus: status` aliases the V2 `VFolder.status`
@@ -841,6 +848,11 @@ const VFolderNodesV2: React.FC<VFolderNodesV2Props> = ({
         vfolderId={deployFallbackVfolderId ?? undefined}
         onClose={() => setDeployFallbackVfolderId(null)}
         onDeployed={() => setDeployFallbackVfolderId(null)}
+        onRequestCreateDeployment={toggleCreateDeployment}
+      />
+      <DeploymentSettingModal
+        open={isCreateDeploymentOpen}
+        onRequestClose={toggleCreateDeployment}
       />
       <HostQuotaModal
         open={isHostQuotaModalOpen}
