+++ resources/i18n/de.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Erstellt von",
     "Current": "Aktuell",
     "CurrentRevision": "Aktuelle Revision",
+    "CurrentRevisionAvailableDescription": "Laden Sie die Konfiguration, um sie als Ausgangspunkt für die neue Revision zu verwenden.",
+    "CurrentRevisionAvailableTitle": "Eine aktuelle Revision ist bereits in Kraft",
+    "CustomMode": "Erweiterter Modus",
     "DeleteDeployment": "Bereitstellung löschen",
     "Deploy": "Bereitstellen",
     "DeployConfirm": "Möchten Sie wirklich Revision #{{revisionNumber}} bereitstellen? Die aktuelle Revision wird ersetzt.",
@@ -953,6 +956,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "Die Endpunkt-URL wurde verschoben. Sie wurden zur neuen Bereitstellungsseite weitergeleitet.",
     "LivenessStatus": "Liveness-Status",
+    "LoadCurrentRevision": "Aktuelle Revision laden",
     "ManageAccessTokens": "Zugriffstoken verwalten",
     "Model": "Modell",
     "ModelDefinitionPath": "Pfad zur Modelldefinitionsdatei",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Beim Speichern wird eine neue Revision erstellt. Fortfahren?",
     "NoCurrentRevisionDeployed": "Keine Revision bereitgestellt — fügen Sie eine Revision hinzu, um diesen Dienst zu aktivieren.",
     "NoDeployments": "Keine Bereitstellungen gefunden.",
+    "NoPresetsAvailable": "Kein Bereitstellungs-Preset verfügbar",
+    "NoPresetsAvailableDescription": "Es gibt kein Preset für die Bereitstellung. Erstellen Sie stattdessen eine neue Bereitstellung.",
+    "NoPresetsAvailableSwitchToCustom": "Es ist kein Preset verfügbar. Wechseln Sie in den benutzerdefinierten Modus, um eine Revision manuell hinzuzufügen.",
     "NumberOfReplicas": "Anzahl der Replikate",
+    "OpenCreateDeploymentModal": "Modal zum Erstellen einer Bereitstellung öffnen",
     "OpenToPublic": "Öffentlich",
     "Overview": "Übersicht",
     "Owner": "Eigentümer",
+    "PresetMode": "Preset-Modus",
     "Private": "Privat",
     "PrivateDeploymentAlertTitle": "Private Bereitstellung — verwenden Sie ein Zugriffstoken, um auf den Endpunkt zuzugreifen.",
     "Project": "Projekt",
@@ -1777,6 +1786,7 @@
     "ClearErrors": "Fehler löschen und erneut versuchen",
     "CloneModelFolderConfirm": "Um den Modelldienst auszuführen, muss der Ordner geklont werden. \nMöchten Sie es als <strong>\"{{folderName}}\"</strong> klonen und den Dienst starten?",
     "Controls": "Kontrolliert",
+    "CreateNewDeploymentWithPreset": "Neue Bereitstellung mit Preset erstellen",
     "CreatedAt": "Erstellt am",
     "CreatingService": "Dienst wird erstellt...",
     "CurrentRevisionTitle": "Aktuelle Revision",
@@ -1815,8 +1825,12 @@
     "Image": "Bild",
     "InitialDelay": "Anfängliche Verzögerung",
     "InitialDelayTooltip": "Sekunden, die nach dem Start des Dienstes gewartet werden, bevor der erste Gesundheitscheck durchgeführt wird.",
+    "Interval": "Intervall (s)",
+    "IntervalTooltip": "Sekunden zwischen aufeinanderfolgenden Gesundheitscheck-Versuchen.",
     "MaxRetries": "Max. Wiederholungsversuche",
     "MaxRetriesTooltip": "Maximale Anzahl von Gesundheitscheck-Versuchen, bevor der Dienst als fehlgeschlagen gilt.",
+    "MaxWaitTime": "Max. Wartezeit (s)",
+    "MaxWaitTimeTooltip": "Maximale Sekunden, die der Dienst hat, um auf einen einzelnen Gesundheitscheck zu antworten, bevor er als Fehler gewertet wird.",
     "ModelAndServingConfiguration": "Modell- & Serving-Konfiguration",
     "ModelDefinition": "Modelldefinition",
     "ModelDefinitionPath": "Pfad der Modelldefinitionsdatei",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "Die Runtime-Variante ist in der Servicedefinition nicht angegeben. Bitte wählen Sie ein Deployment-Preset aus, um fortzufahren.",
     "RuntimeVariantSelectionRequired": "Mehrere Runtime-Varianten sind verfügbar. Bitte wählen Sie ein Deployment-Preset aus, um fortzufahren.",
     "SeeContainerLogs": "Siehe Containerprotokolle",
-    "SelectDeploymentPreset": "Deployment-Preset auswählen",
     "ServiceAlreadyExists": "Servicename existiert bereits",
     "ServiceConfigurationNotValid": "Die Dienstkonfiguration ist ungültig. \nBitte wenden Sie sich an den Administrator.",
     "ServiceCreated": "Der Modelldienst {{ name }} wurde erfolgreich erstellt.",
@@ -3011,7 +3024,6 @@
     "Copied": "Kopiert",
     "Copy": "Kopieren"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Eine Sitzung erstellen",
     "CreateFolder": "Erstellen Sie einen neuen Speicherordner",
