+++ resources/i18n/el.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Δημιουργήθηκε από",
     "Current": "Τρέχουσα",
     "CurrentRevision": "Τρέχουσα Revision",
+    "CurrentRevisionAvailableDescription": "Φορτώστε τη διαμόρφωσή της για να τη χρησιμοποιήσετε ως σημείο εκκίνησης για τη νέα Revision.",
+    "CurrentRevisionAvailableTitle": "Μια τρέχουσα Revision βρίσκεται ήδη σε ισχύ",
+    "CustomMode": "Σύνθετη Λειτουργία",
     "DeleteDeployment": "Διαγραφή ανάπτυξης",
     "Deploy": "Ανάπτυξη",
     "DeployConfirm": "Είστε σίγουροι ότι θέλετε να αναπτύξετε τη Revision #{{revisionNumber}}; Η τρέχουσα Revision θα αντικατασταθεί.",
@@ -953,6 +956,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "Το URL τελικού σημείου μετακινήθηκε. Ανακατευθυνθήκατε στη νέα σελίδα αναπτύξεων.",
     "LivenessStatus": "Κατάσταση Liveness",
+    "LoadCurrentRevision": "Φόρτωση τρέχουσας Revision",
     "ManageAccessTokens": "Διαχείριση διακριτικών πρόσβασης",
     "Model": "Μοντέλο",
     "ModelDefinitionPath": "Διαδρομή αρχείου ορισμού μοντέλου",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Η αποθήκευση θα δημιουργήσει νέα Revision. Συνέχεια;",
     "NoCurrentRevisionDeployed": "Δεν υπάρχει αναπτυγμένη αναθεώρηση — προσθέστε μια αναθεώρηση για να ενεργοποιήσετε αυτή την υπηρεσία.",
     "NoDeployments": "Δεν βρέθηκαν αναπτύξεις.",
+    "NoPresetsAvailable": "Δεν υπάρχει διαθέσιμο προκαθορισμένο ανάπτυξης",
+    "NoPresetsAvailableDescription": "Δεν υπάρχει προκαθορισμένο για ανάπτυξη. Δημιουργήστε μια νέα ανάπτυξη.",
+    "NoPresetsAvailableSwitchToCustom": "Δεν υπάρχει διαθέσιμο προκαθορισμένο. Μεταβείτε σε προσαρμοσμένη λειτουργία για να προσθέσετε Revision χειροκίνητα.",
     "NumberOfReplicas": "Αριθμός αντιγράφων",
+    "OpenCreateDeploymentModal": "Άνοιγμα αναδυόμενου παραθύρου δημιουργίας ανάπτυξης",
     "OpenToPublic": "Δημόσιο",
     "Overview": "Επισκόπηση",
     "Owner": "Ιδιοκτήτης",
+    "PresetMode": "Λειτουργία Προκαθορισμένου",
     "Private": "Ιδιωτικό",
     "PrivateDeploymentAlertTitle": "Ιδιωτική ανάπτυξη — χρησιμοποιήστε ένα διακριτικό πρόσβασης για να αποκτήσετε πρόσβαση στο τελικό σημείο.",
     "Project": "Έργο",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Εκκαθάριση σφάλματος και επανάληψη",
     "CloneModelFolderConfirm": "Για να εκτελέσετε την υπηρεσία μοντέλου, ο φάκελος πρέπει να κλωνοποιηθεί. \nΘα θέλατε να το κλωνοποιήσετε ως <strong>\"{{folderName}}\"</strong> και να ξεκινήσετε την υπηρεσία;",
     "Controls": "Έλεγχοι",
+    "CreateNewDeploymentWithPreset": "Δημιουργία νέου Deployment με Preset",
     "CreatedAt": "Δημιουργήθηκε στο",
     "CreatingService": "Δημιουργία υπηρεσίας...",
     "CurrentRevisionTitle": "Τρέχουσα αναθεώρηση",
@@ -1813,8 +1823,12 @@
     "Image": "Εικόνα",
     "InitialDelay": "Αρχική καθυστέρηση",
     "InitialDelayTooltip": "Δευτερόλεπτα αναμονής πριν από τον πρώτο έλεγχο υγείας μετά την εκκίνηση της υπηρεσίας.",
+    "Interval": "Διάστημα (s)",
+    "IntervalTooltip": "Δευτερόλεπτα μεταξύ διαδοχικών προσπαθειών ελέγχου υγείας.",
     "MaxRetries": "Μέγιστες επαναλήψεις",
     "MaxRetriesTooltip": "Μέγιστος αριθμός προσπαθειών ελέγχου υγείας πριν η υπηρεσία θεωρηθεί αποτυχημένη.",
+    "MaxWaitTime": "Μέγιστος χρόνος αναμονής (s)",
+    "MaxWaitTimeTooltip": "Μέγιστος αριθμός δευτερολέπτων που έχει η υπηρεσία για να ανταποκριθεί σε έναν μεμονωμένο έλεγχο υγείας πριν θεωρηθεί αποτυχία.",
     "ModelAndServingConfiguration": "Διαμόρφωση μοντέλου & σερβιρίσματος",
     "ModelDefinition": "Ορισμός μοντέλου",
     "ModelDefinitionPath": "Διαδρομή αρχείου ορισμού μοντέλου",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "Η παραλλαγή χρόνου εκτέλεσης δεν έχει οριστεί στον ορισμό της υπηρεσίας. Επιλέξτε ένα προκαθορισμένο deployment για να συνεχίσετε.",
     "RuntimeVariantSelectionRequired": "Διατίθενται πολλές παραλλαγές χρόνου εκτέλεσης. Επιλέξτε ένα προκαθορισμένο deployment για να συνεχίσετε.",
     "SeeContainerLogs": "Βλέπε αρχεία καταγραφής κοντέινερ",
-    "SelectDeploymentPreset": "Επιλογή προκαθορισμένου Deployment",
     "ServiceAlreadyExists": "Το όνομα της υπηρεσίας υπάρχει ήδη",
     "ServiceConfigurationNotValid": "Η διαμόρφωση της υπηρεσίας δεν είναι έγκυρη. \nΕπικοινωνήστε με τον διαχειριστή.",
     "ServiceCreated": "Η υπηρεσία μοντέλου {{ όνομα }} δημιουργήθηκε με επιτυχία.",
@@ -3009,7 +3022,6 @@
     "Copied": "Αντιγράφηκε",
     "Copy": "αντίγραφο"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Δημιουργία συνεδρίας",
     "CreateFolder": "Δημιουργήστε έναν νέο φάκελο αποθήκευσης",
