+++ resources/i18n/en.json
@@ -936,6 +936,9 @@
     "CreatedBy": "Created By",
     "Current": "Current",
     "CurrentRevision": "Current Revision",
+    "CurrentRevisionAvailableDescription": "Load its configuration to use it as a starting point for the new revision.",
+    "CurrentRevisionAvailableTitle": "A current revision is already in effect",
+    "CustomMode": "Advanced Mode",
     "DeleteDeployment": "Delete Deployment",
     "Deploy": "Deploy",
     "DeployConfirm": "Are you sure you want to deploy revision #{{revisionNumber}}? The current revision will be replaced.",
@@ -972,6 +975,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "The endpoint URL has moved. You have been redirected to the new Deployments page.",
     "LivenessStatus": "Liveness Status",
+    "LoadCurrentRevision": "Load current revision",
     "ManageAccessTokens": "Manage Access Tokens",
     "Model": "Model",
     "ModelDefinitionPath": "Model Definition File Path",
@@ -987,10 +991,15 @@
     "NewRevisionWillBeCreatedConfirm": "Saving will create a new Revision. Continue?",
     "NoCurrentRevisionDeployed": "No revision is deployed — add a revision to activate this service.",
     "NoDeployments": "No deployments found.",
+    "NoPresetsAvailable": "No deployment preset available",
+    "NoPresetsAvailableDescription": "There is no preset to deploy with. Create a new deployment instead.",
+    "NoPresetsAvailableSwitchToCustom": "There is no preset available. Switch to Custom mode to add a revision manually.",
     "NumberOfReplicas": "Number of Replicas",
+    "OpenCreateDeploymentModal": "Open create deployment modal",
     "OpenToPublic": "Open To Public",
     "Overview": "Overview",
     "Owner": "Owner",
+    "PresetMode": "Preset Mode",
     "Private": "Private",
     "PrivateDeploymentAlertTitle": "Private deployment — use an access token to access the endpoint.",
     "Project": "Project",
@@ -1794,6 +1803,7 @@
     "ClearErrors": "Clear error and retry",
     "CloneModelFolderConfirm": "To run the model service, the folder needs to be cloned. Would you like to clone it as <strong>\"{{folderName}}\"</strong> and start the service?",
     "Controls": "Controls",
+    "CreateNewDeploymentWithPreset": "Create New Deployment with Preset",
     "CreatedAt": "Created At",
     "CreatingService": "Creating service...",
     "CurrentRevisionTitle": "Current Revision",
@@ -1832,8 +1842,12 @@
     "Image": "Image",
     "InitialDelay": "Initial Delay",
     "InitialDelayTooltip": "Seconds to wait before the first health check after the service starts.",
+    "Interval": "Interval (s)",
+    "IntervalTooltip": "Seconds between consecutive health check attempts.",
     "MaxRetries": "Max Retries",
     "MaxRetriesTooltip": "Maximum number of health check attempts before the service is considered failed.",
+    "MaxWaitTime": "Max Wait Time (s)",
+    "MaxWaitTimeTooltip": "Maximum seconds the service has to respond to a single health check before it is treated as a failure.",
     "ModelAndServingConfiguration": "Model & Serving Configuration",
     "ModelDefinition": "Model Definition",
     "ModelDefinitionPath": "Model Definition File Path",
@@ -1920,7 +1934,6 @@
     "RuntimeVariantNotSpecified": "Runtime variant is not specified in the service definition. Please select a deployment preset to continue.",
     "RuntimeVariantSelectionRequired": "Multiple runtime variants are available. Please select a deployment preset to continue.",
     "SeeContainerLogs": "See Container Logs",
-    "SelectDeploymentPreset": "Select Deployment Preset",
     "ServiceAlreadyExists": "Service name already exists",
     "ServiceConfigurationNotValid": "The service configuration is invalid. Please contact the administrator.",
     "ServiceCreated": "Model service {{ name }} has been created successfully.",
@@ -3037,7 +3050,6 @@
     "Copied": "Copied",
     "Copy": "Copy"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Create a Session",
     "CreateFolder": "Create New Storage Folder",
