+++ resources/i18n/es.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Creado por",
     "Current": "Actual",
     "CurrentRevision": "Revision Actual",
+    "CurrentRevisionAvailableDescription": "Cargue su configuración para usarla como punto de partida para la nueva Revision.",
+    "CurrentRevisionAvailableTitle": "Ya hay una Revision actual en vigor",
+    "CustomMode": "Modo Avanzado",
     "DeleteDeployment": "Eliminar despliegue",
     "Deploy": "Desplegar",
     "DeployConfirm": "¿Está seguro de que desea desplegar la Revision #{{revisionNumber}}? La Revision actual será reemplazada.",
@@ -953,6 +956,7 @@
     "Image": "Imagen",
     "LegacyRouteRedirected": "La URL del endpoint ha cambiado. Ha sido redirigido a la nueva página de despliegues.",
     "LivenessStatus": "Estado de Liveness",
+    "LoadCurrentRevision": "Cargar Revision actual",
     "ManageAccessTokens": "Administrar tokens de acceso",
     "Model": "Modelo",
     "ModelDefinitionPath": "Ruta del archivo de definición del modelo",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Guardar creará una nueva Revision. ¿Continuar?",
     "NoCurrentRevisionDeployed": "No hay ninguna revisión desplegada — agregue una revisión para activar este servicio.",
     "NoDeployments": "No se encontraron despliegues.",
+    "NoPresetsAvailable": "No hay preset de despliegue disponible",
+    "NoPresetsAvailableDescription": "No hay ningún preset con el que desplegar. Cree un nuevo despliegue en su lugar.",
+    "NoPresetsAvailableSwitchToCustom": "No hay preset disponible. Cambie al modo Personalizado para agregar una Revision manualmente.",
     "NumberOfReplicas": "Número de réplicas",
+    "OpenCreateDeploymentModal": "Abrir modal de creación de despliegue",
     "OpenToPublic": "Público",
     "Overview": "Resumen",
     "Owner": "Propietario",
+    "PresetMode": "Modo Preset",
     "Private": "Privado",
     "PrivateDeploymentAlertTitle": "Despliegue privado — use un token de acceso para acceder al endpoint.",
     "Project": "Proyecto",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Borrar error y reintentar",
     "CloneModelFolderConfirm": "Para ejecutar el servicio modelo, la carpeta debe ser clonada. \n¿Le gustaría clonarlo como <strong>\"{{folderName}}\"</strong> e iniciar el servicio?",
     "Controls": "Controla",
+    "CreateNewDeploymentWithPreset": "Crear nuevo despliegue con preset",
     "CreatedAt": "Creado en",
     "CreatingService": "Creando servicio...",
     "CurrentRevisionTitle": "Revisión actual",
@@ -1813,8 +1823,12 @@
     "Image": "Imagen",
     "InitialDelay": "Retraso inicial",
     "InitialDelayTooltip": "Segundos a esperar antes del primer chequeo de salud tras iniciar el servicio.",
+    "Interval": "Intervalo (s)",
+    "IntervalTooltip": "Segundos entre intentos consecutivos de chequeo de salud.",
     "MaxRetries": "Reintentos máximos",
     "MaxRetriesTooltip": "Número máximo de intentos de chequeo de salud antes de considerar que el servicio ha fallado.",
+    "MaxWaitTime": "Tiempo máx. de espera (s)",
+    "MaxWaitTimeTooltip": "Segundos máximos que tiene el servicio para responder a un único chequeo de salud antes de considerarse un fallo.",
     "ModelAndServingConfiguration": "Configuración de modelo y servicio",
     "ModelDefinition": "Definición de modelo",
     "ModelDefinitionPath": "Ruta del archivo de definición del modelo",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "La variante de runtime no está especificada en la definición del servicio. Seleccione un preset de despliegue para continuar.",
     "RuntimeVariantSelectionRequired": "Hay varias variantes de runtime disponibles. Seleccione un preset de despliegue para continuar.",
     "SeeContainerLogs": "Ver registros de contenedores",
-    "SelectDeploymentPreset": "Seleccionar preset de despliegue",
     "ServiceAlreadyExists": "El nombre del servicio ya existe",
     "ServiceConfigurationNotValid": "La configuración del servicio no es válida. \nPóngase en contacto con el administrador.",
     "ServiceCreated": "El servicio modelo {{ name }} se ha creado correctamente.",
@@ -3009,7 +3022,6 @@
     "Copied": "Copiado",
     "Copy": "Copiar"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Crear una sesión",
     "CreateFolder": "Crear una nueva carpeta de almacenamiento",
