+++ resources/i18n/fi.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Luonut",
     "Current": "Nykyinen",
     "CurrentRevision": "Nykyinen Revision",
+    "CurrentRevisionAvailableDescription": "Lataa sen konfiguraatio uuden Revision lähtökohdaksi.",
+    "CurrentRevisionAvailableTitle": "Nykyinen Revision on jo voimassa",
+    "CustomMode": "Edistynyt tila",
     "DeleteDeployment": "Poista käyttöönotto",
     "Deploy": "Ota käyttöön",
     "DeployConfirm": "Haluatko varmasti ottaa käyttöön Revision #{{revisionNumber}}? Nykyinen Revision korvataan.",
@@ -953,6 +956,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "Päätepisteen URL on muuttunut. Sinut on ohjattu uudelle käyttöönottosivulle.",
     "LivenessStatus": "Liveness-tila",
+    "LoadCurrentRevision": "Lataa nykyinen Revision",
     "ManageAccessTokens": "Hallitse käyttöoikeustunnuksia",
     "Model": "Malli",
     "ModelDefinitionPath": "Mallimäärittelytiedoston polku",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Tallennus luo uuden Revision. Jatketaanko?",
     "NoCurrentRevisionDeployed": "Versiota ei ole otettu käyttöön — lisää versio aktivoidaksesi tämän palvelun.",
     "NoDeployments": "Käyttöönottoja ei löydy.",
+    "NoPresetsAvailable": "Ei käyttöönottomalleja saatavilla",
+    "NoPresetsAvailableDescription": "Käyttöönottoon ei ole saatavilla mallia. Luo uusi käyttöönotto sen sijaan.",
+    "NoPresetsAvailableSwitchToCustom": "Malleja ei ole saatavilla. Vaihda mukautettuun tilaan lisätäksesi Revision manuaalisesti.",
     "NumberOfReplicas": "Replikoiden määrä",
+    "OpenCreateDeploymentModal": "Avaa käyttöönoton luontimodaali",
     "OpenToPublic": "Julkinen",
     "Overview": "Yleiskatsaus",
     "Owner": "Omistaja",
+    "PresetMode": "Mallitila",
     "Private": "Yksityinen",
     "PrivateDeploymentAlertTitle": "Yksityinen käyttöönotto — käytä käyttöoikeustunnusta päätepisteen käyttämiseen.",
     "Project": "Projekti",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Tyhjennä virhe ja yritä uudelleen",
     "CloneModelFolderConfirm": "Mallipalvelun suorittamiseksi kansio on kloonata. \nHaluatko kloonata sen <strong>\"{{folderName}}\"</strong> ja aloittaa palvelun?",
     "Controls": "Ohjaimet",
+    "CreateNewDeploymentWithPreset": "Luo uusi käyttöönotto esiasetuksella",
     "CreatedAt": "Created at",
     "CreatingService": "Creating service...",
     "CurrentRevisionTitle": "Nykyinen revisio",
@@ -1813,8 +1823,12 @@
     "Image": "Kuva",
     "InitialDelay": "Alkuviive",
     "InitialDelayTooltip": "Sekunteja odottaa ennen ensimmäistä terveystarkastusta palvelun käynnistymisen jälkeen.",
+    "Interval": "Aikaväli (s)",
+    "IntervalTooltip": "Sekuntia peräkkäisten terveystarkastusyritysten välillä.",
     "MaxRetries": "Max Uudelleenyritykset",
     "MaxRetriesTooltip": "Enimmäismäärä terveystarkastusyrityksiä ennen kuin palvelu katsotaan epäonnistuneeksi.",
+    "MaxWaitTime": "Enimmäisodotusaika (s)",
+    "MaxWaitTimeTooltip": "Sekuntien enimmäismäärä, jonka palvelulla on aikaa vastata yksittäiseen terveystarkastukseen ennen kuin se katsotaan epäonnistuneeksi.",
     "ModelAndServingConfiguration": "Malli- ja palvelukonfiguraatio",
     "ModelDefinition": "Mallin määritelmä",
     "ModelDefinitionPath": "Mallin määritelmän tiedostopolku",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "Runtime-varianttia ei ole määritetty palvelumäärittelyssä. Valitse käyttöönoton esiasetus jatkaaksesi.",
     "RuntimeVariantSelectionRequired": "Useita runtime-variantteja on saatavilla. Valitse käyttöönoton esiasetus jatkaaksesi.",
     "SeeContainerLogs": "Katso Säilön lokit",
-    "SelectDeploymentPreset": "Valitse käyttöönoton esiasetus",
     "ServiceAlreadyExists": "Palvelun nimi on jo olemassa",
     "ServiceConfigurationNotValid": "Palvelun kokoonpano on virheellinen. \nOta yhteyttä järjestelmänvalvojaan.",
     "ServiceCreated": "Mallipalvelun {{ name }} luominen onnistui.",
@@ -3009,7 +3022,6 @@
     "Copied": "Kopioitu",
     "Copy": "Kopio"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Luo istunto",
     "CreateFolder": "Luo uusi tallennuskansio",
