+++ resources/i18n/fr.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Créé par",
     "Current": "Actuelle",
     "CurrentRevision": "Revision Actuelle",
+    "CurrentRevisionAvailableDescription": "Chargez sa configuration pour l'utiliser comme point de départ de la nouvelle Revision.",
+    "CurrentRevisionAvailableTitle": "Une Revision actuelle est déjà en vigueur",
+    "CustomMode": "Mode Avancé",
     "DeleteDeployment": "Supprimer le déploiement",
     "Deploy": "Déployer",
     "DeployConfirm": "Êtes-vous sûr de vouloir déployer la Revision #{{revisionNumber}} ? La Revision actuelle sera remplacée.",
@@ -953,6 +956,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "L'URL du point de terminaison a changé. Vous avez été redirigé vers la nouvelle page de déploiements.",
     "LivenessStatus": "État Liveness",
+    "LoadCurrentRevision": "Charger la Revision actuelle",
     "ManageAccessTokens": "Gérer les jetons d'accès",
     "Model": "Modèle",
     "ModelDefinitionPath": "Chemin du fichier de définition du modèle",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "L'enregistrement créera une nouvelle Revision. Continuer ?",
     "NoCurrentRevisionDeployed": "Aucune révision n'est déployée — ajoutez une révision pour activer ce service.",
     "NoDeployments": "Aucun déploiement trouvé.",
+    "NoPresetsAvailable": "Aucun preset de déploiement disponible",
+    "NoPresetsAvailableDescription": "Il n'existe aucun preset pour déployer. Créez plutôt un nouveau déploiement.",
+    "NoPresetsAvailableSwitchToCustom": "Aucun preset disponible. Passez en mode Personnalisé pour ajouter une Revision manuellement.",
     "NumberOfReplicas": "Nombre de répliques",
+    "OpenCreateDeploymentModal": "Ouvrir le modal de création de déploiement",
     "OpenToPublic": "Public",
     "Overview": "Aperçu",
     "Owner": "Propriétaire",
+    "PresetMode": "Mode Preset",
     "Private": "Privé",
     "PrivateDeploymentAlertTitle": "Déploiement privé — utilisez un jeton d'accès pour accéder au point de terminaison.",
     "Project": "Projet",
@@ -1776,6 +1785,7 @@
     "ClearErrors": "Effacer l'erreur et réessayer",
     "CloneModelFolderConfirm": "Pour exécuter le service modèle, le dossier doit être cloné. \nSouhaitez-vous le cloner comme <strong>\"{{folderName}}\"</strong> et démarrer le service?",
     "Controls": "Contrôles",
+    "CreateNewDeploymentWithPreset": "Créer un nouveau déploiement avec préréglage",
     "CreatedAt": "Créé à",
     "CreatingService": "Création du service...",
     "CurrentRevisionTitle": "Révision actuelle",
@@ -1814,8 +1824,12 @@
     "Image": "Image",
     "InitialDelay": "Délai initial",
     "InitialDelayTooltip": "Secondes à attendre avant le premier contrôle de santé après le démarrage du service.",
+    "Interval": "Intervalle (s)",
+    "IntervalTooltip": "Secondes entre les tentatives consécutives de contrôle de santé.",
     "MaxRetries": "Nombre maximal de tentatives",
     "MaxRetriesTooltip": "Nombre maximum de tentatives de contrôle de santé avant que le service soit considéré comme défaillant.",
+    "MaxWaitTime": "Temps d'attente max. (s)",
+    "MaxWaitTimeTooltip": "Secondes maximum dont dispose le service pour répondre à un contrôle de santé unique avant qu'il soit considéré comme un échec.",
     "ModelAndServingConfiguration": "Configuration du modèle et du service",
     "ModelDefinition": "Définition du modèle",
     "ModelDefinitionPath": "Chemin du fichier de définition du modèle",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "La variante de runtime n'est pas spécifiée dans la définition du service. Veuillez sélectionner un préréglage de déploiement pour continuer.",
     "RuntimeVariantSelectionRequired": "Plusieurs variantes de runtime sont disponibles. Veuillez sélectionner un préréglage de déploiement pour continuer.",
     "SeeContainerLogs": "Voir les journaux de conteneur",
-    "SelectDeploymentPreset": "Sélectionner un préréglage de déploiement",
     "ServiceAlreadyExists": "Le nom du service existe déjà",
     "ServiceConfigurationNotValid": "La configuration du service n'est pas valide. \nVeuillez contacter l'administrateur.",
     "ServiceCreated": "Le service modèle {{ name }} a été créé avec succès.",
@@ -3012,7 +3025,6 @@
     "Copied": "Copié",
     "Copy": "Copie"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Créer une session",
     "CreateFolder": "Créer un nouveau dossier de stockage",
