+++ resources/i18n/id.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Dibuat Oleh",
     "Current": "Saat Ini",
     "CurrentRevision": "Revision Saat Ini",
+    "CurrentRevisionAvailableDescription": "Muat konfigurasinya untuk digunakan sebagai titik awal revision baru.",
+    "CurrentRevisionAvailableTitle": "Revision saat ini sudah berlaku",
+    "CustomMode": "Mode Lanjutan",
     "DeleteDeployment": "Hapus Penerapan",
     "Deploy": "Terapkan",
     "DeployConfirm": "Apakah Anda yakin ingin menerapkan Revision #{{revisionNumber}}? Revision saat ini akan diganti.",
@@ -953,6 +956,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "URL titik akhir telah dipindahkan. Anda telah diarahkan ke halaman Penerapan baru.",
     "LivenessStatus": "Status Liveness",
+    "LoadCurrentRevision": "Muat Revision saat ini",
     "ManageAccessTokens": "Kelola token akses",
     "Model": "Model",
     "ModelDefinitionPath": "Jalur File Definisi Model",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Menyimpan akan membuat Revision baru. Lanjutkan?",
     "NoCurrentRevisionDeployed": "Tidak ada revisi yang diterapkan — tambahkan revisi untuk mengaktifkan layanan ini.",
     "NoDeployments": "Tidak ada penerapan ditemukan.",
+    "NoPresetsAvailable": "Tidak ada preset penerapan tersedia",
+    "NoPresetsAvailableDescription": "Tidak ada preset untuk melakukan penerapan. Buat penerapan baru sebagai gantinya.",
+    "NoPresetsAvailableSwitchToCustom": "Tidak ada preset yang tersedia. Alihkan ke mode Kustom untuk menambahkan revision secara manual.",
     "NumberOfReplicas": "Jumlah Replika",
+    "OpenCreateDeploymentModal": "Buka modal buat penerapan",
     "OpenToPublic": "Terbuka untuk Publik",
     "Overview": "Ikhtisar",
     "Owner": "Pemilik",
+    "PresetMode": "Mode Preset",
     "Private": "Privat",
     "PrivateDeploymentAlertTitle": "Penerapan privat — gunakan token akses untuk mengakses endpoint.",
     "Project": "Proyek",
@@ -1778,6 +1787,7 @@
     "ClearErrors": "Hapus kesalahan dan coba lagi",
     "CloneModelFolderConfirm": "Untuk menjalankan layanan model, folder perlu dikloning. \nApakah Anda ingin mengkloningnya sebagai <strong>\"{{folderName}}\"</strong> dan memulai layanan?",
     "Controls": "Kontrol",
+    "CreateNewDeploymentWithPreset": "Buat Deployment Baru dengan Preset",
     "CreatedAt": "Dibuat di",
     "CreatingService": "Membuat layanan...",
     "CurrentRevisionTitle": "Revisi Saat Ini",
@@ -1816,8 +1826,12 @@
     "Image": "Gambar",
     "InitialDelay": "Penundaan Awal",
     "InitialDelayTooltip": "Detik untuk menunggu sebelum pemeriksaan kesehatan pertama setelah layanan dimulai.",
+    "Interval": "Interval (s)",
+    "IntervalTooltip": "Detik antara percobaan pemeriksaan kesehatan yang berurutan.",
     "MaxRetries": "Percobaan Ulang Maks",
     "MaxRetriesTooltip": "Jumlah maksimum percobaan pemeriksaan kesehatan sebelum layanan dianggap gagal.",
+    "MaxWaitTime": "Waktu Tunggu Maks (s)",
+    "MaxWaitTimeTooltip": "Detik maksimum yang dimiliki layanan untuk merespons satu pemeriksaan kesehatan sebelum dianggap gagal.",
     "ModelAndServingConfiguration": "Konfigurasi Model & Serving",
     "ModelDefinition": "Definisi Model",
     "ModelDefinitionPath": "Jalur File Definisi Model",
@@ -1904,7 +1918,6 @@
     "RuntimeVariantNotSpecified": "Varian runtime tidak ditentukan dalam definisi layanan. Silakan pilih preset deployment untuk melanjutkan.",
     "RuntimeVariantSelectionRequired": "Beberapa varian runtime tersedia. Silakan pilih preset deployment untuk melanjutkan.",
     "SeeContainerLogs": "Lihat Log Kontainer",
-    "SelectDeploymentPreset": "Pilih Preset Deployment",
     "ServiceAlreadyExists": "Nama layanan sudah ada",
     "ServiceConfigurationNotValid": "Konfigurasi layanan tidak valid. \nSilakan hubungi administrator.",
     "ServiceCreated": "Layanan model {{ name }} telah berhasil dibuat.",
@@ -3013,7 +3026,6 @@
     "Copied": "Disalin",
     "Copy": "Menyalin"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Membuat Sesi",
     "CreateFolder": "Buat Folder Penyimpanan Baru",
