+++ resources/i18n/it.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Creato da",
     "Current": "Corrente",
     "CurrentRevision": "Revision Corrente",
+    "CurrentRevisionAvailableDescription": "Carica la sua configurazione per usarla come punto di partenza per la nuova Revision.",
+    "CurrentRevisionAvailableTitle": "Una Revision corrente è già in vigore",
+    "CustomMode": "Modalità Avanzata",
     "DeleteDeployment": "Elimina distribuzione",
     "Deploy": "Distribuisci",
     "DeployConfirm": "Sei sicuro di voler distribuire la Revision #{{revisionNumber}}? La Revision attuale verrà sostituita.",
@@ -953,6 +956,7 @@
     "Image": "Image",
     "LegacyRouteRedirected": "L'URL dell'endpoint è stata spostata. Sei stato reindirizzato alla nuova pagina Distribuzioni.",
     "LivenessStatus": "Stato Liveness",
+    "LoadCurrentRevision": "Carica Revision corrente",
     "ManageAccessTokens": "Gestisci token di accesso",
     "Model": "Modello",
     "ModelDefinitionPath": "Percorso file di definizione modello",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Il salvataggio creerà una nuova Revision. Continuare?",
     "NoCurrentRevisionDeployed": "Nessuna revisione è distribuita — aggiungi una revisione per attivare questo servizio.",
     "NoDeployments": "Nessuna distribuzione trovata.",
+    "NoPresetsAvailable": "Nessun preset di distribuzione disponibile",
+    "NoPresetsAvailableDescription": "Non c'è nessun preset con cui distribuire. Crea invece una nuova distribuzione.",
+    "NoPresetsAvailableSwitchToCustom": "Nessun preset disponibile. Passa alla modalità Personalizzata per aggiungere una Revision manualmente.",
     "NumberOfReplicas": "Numero di repliche",
+    "OpenCreateDeploymentModal": "Apri modale di creazione distribuzione",
     "OpenToPublic": "Pubblico",
     "Overview": "Panoramica",
     "Owner": "Proprietario",
+    "PresetMode": "Modalità Preset",
     "Private": "Privato",
     "PrivateDeploymentAlertTitle": "Distribuzione privata — usa un token di accesso per accedere all'endpoint.",
     "Project": "Progetto",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Cancellare l'errore e riprovare",
     "CloneModelFolderConfirm": "Per eseguire il servizio modello, la cartella deve essere clonata. \nVorresti clonarlo come <strong>\"{{folderName}}\"</strong> e avviare il servizio?",
     "Controls": "Controlli",
+    "CreateNewDeploymentWithPreset": "Crea nuova distribuzione con preset",
     "CreatedAt": "Creato a",
     "CreatingService": "Creazione del servizio...",
     "CurrentRevisionTitle": "Revisione corrente",
@@ -1813,8 +1823,12 @@
     "Image": "Immagine",
     "InitialDelay": "Ritardo iniziale",
     "InitialDelayTooltip": "Secondi di attesa prima del primo controllo di integrità dopo l'avvio del servizio.",
+    "Interval": "Intervallo (s)",
+    "IntervalTooltip": "Secondi tra i tentativi consecutivi di controllo di integrità.",
     "MaxRetries": "Numero massimo di tentativi",
     "MaxRetriesTooltip": "Numero massimo di tentativi di controllo di integrità prima che il servizio venga considerato non funzionante.",
+    "MaxWaitTime": "Tempo di attesa massimo (s)",
+    "MaxWaitTimeTooltip": "Secondi massimi a disposizione del servizio per rispondere a un singolo controllo di integrità prima che venga considerato un errore.",
     "ModelAndServingConfiguration": "Configurazione modello e serving",
     "ModelDefinition": "Definizione del modello",
     "ModelDefinitionPath": "Percorso del file di definizione del modello",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "La variante di runtime non è specificata nella definizione del servizio. Selezionare un preset di distribuzione per continuare.",
     "RuntimeVariantSelectionRequired": "Sono disponibili più varianti di runtime. Selezionare un preset di distribuzione per continuare.",
     "SeeContainerLogs": "Consulta Log del contenitore",
-    "SelectDeploymentPreset": "Seleziona preset di distribuzione",
     "ServiceAlreadyExists": "Il nome del servizio esiste già",
     "ServiceConfigurationNotValid": "La configurazione del servizio non è valida. \nSi prega di contattare l'amministratore.",
     "ServiceCreated": "Il servizio modello {{ name }} è stato creato correttamente.",
@@ -3009,7 +3022,6 @@
     "Copied": "Copiato",
     "Copy": "copia"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Creare una sessione",
     "CreateFolder": "Crea una nuova cartella di archiviazione",
