+++ resources/i18n/ja.json
@@ -918,6 +918,9 @@
     "CreatedBy": "作成者",
     "Current": "現在",
     "CurrentRevision": "現在の Revision",
+    "CurrentRevisionAvailableDescription": "その設定を読み込んで、新しい Revision の開始点として使用できます。",
+    "CurrentRevisionAvailableTitle": "現在の Revision がすでに有効です",
+    "CustomMode": "高度な設定モード",
     "DeleteDeployment": "デプロイを削除",
     "Deploy": "デプロイ",
     "DeployConfirm": "Revision #{{revisionNumber}} をデプロイしてもよろしいですか？現在の Revision は置き換えられます。",
@@ -953,6 +956,7 @@
     "Image": "イメージ",
     "LegacyRouteRedirected": "エンドポイントURLが変更されました。新しいデプロイページにリダイレクトされました。",
     "LivenessStatus": "Liveness 状態",
+    "LoadCurrentRevision": "現在の Revision を読み込む",
     "ManageAccessTokens": "アクセストークンを管理",
     "Model": "モデル",
     "ModelDefinitionPath": "モデル定義ファイルパス",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "保存すると新しい Revision が作成されます。続行しますか？",
     "NoCurrentRevisionDeployed": "デプロイされたリビジョンがありません。リビジョンを追加してこのサービスを有効化してください。",
     "NoDeployments": "デプロイが見つかりません。",
+    "NoPresetsAvailable": "利用可能なデプロイプリセットがありません",
+    "NoPresetsAvailableDescription": "デプロイに使用できるプリセットがありません。代わりに新しいデプロイを作成してください。",
+    "NoPresetsAvailableSwitchToCustom": "利用可能なプリセットがありません。カスタムモードに切り替えて、手動で Revision を追加してください。",
     "NumberOfReplicas": "レプリカ数",
+    "OpenCreateDeploymentModal": "デプロイ作成モーダルを開く",
     "OpenToPublic": "公開",
     "Overview": "概要",
     "Owner": "オーナー",
+    "PresetMode": "プリセットモード",
     "Private": "非公開",
     "PrivateDeploymentAlertTitle": "非公開デプロイメントです。エンドポイントの使用にはアクセストークンが必要です。",
     "Project": "プロジェクト",
@@ -1777,6 +1786,7 @@
     "ClearErrors": "エラーをクリアして再試行",
     "CloneModelFolderConfirm": "モデルサービスを実行するには、フォルダーをクローン化する必要があります。 \n<strong>\"{{folderName}}\"</strong>としてクローンしてサービスを開始しますか？",
     "Controls": "操作",
+    "CreateNewDeploymentWithPreset": "プリセットで新規デプロイメント作成",
     "CreatedAt": "作成時間",
     "CreatingService": "サービスを作成中...",
     "CurrentRevisionTitle": "現在のリビジョン",
@@ -1815,8 +1825,12 @@
     "Image": "イメージ",
     "InitialDelay": "初期遅延",
     "InitialDelayTooltip": "サービス起動後、最初のヘルスチェックを行うまで待機する秒数です。",
+    "Interval": "インターバル (s)",
+    "IntervalTooltip": "連続するヘルスチェックの試行間隔（秒）です。",
     "MaxRetries": "最大再試行回数",
     "MaxRetriesTooltip": "サービスが失敗と判断されるまでのヘルスチェックの最大試行回数です。",
+    "MaxWaitTime": "最大待機時間 (s)",
+    "MaxWaitTimeTooltip": "1回のヘルスチェックに対してサービスが応答するための最大秒数です。この時間を超えると失敗と判断されます。",
     "ModelAndServingConfiguration": "モデル & サービング設定",
     "ModelDefinition": "モデル定義",
     "ModelDefinitionPath": "モデル定義ファイルのパス",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "サービス定義にランタイムバリアントが指定されていません。続行するにはデプロイメントプリセットを選択してください。",
     "RuntimeVariantSelectionRequired": "複数のランタイムバリアントが利用可能です。続行するにはデプロイメントプリセットを選択してください。",
     "SeeContainerLogs": "コンテナログを参照",
-    "SelectDeploymentPreset": "デプロイメントプリセットを選択",
     "ServiceAlreadyExists": "サービス名はすでに存在します",
     "ServiceConfigurationNotValid": "サービス構成は無効です。\n管理者に連絡してください。",
     "ServiceCreated": "モデル サービス {{ name }} が正常に作成されました。",
@@ -3011,7 +3024,6 @@
     "Copied": "コピーされました",
     "Copy": "コピー"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "セッションの作成",
     "CreateFolder": "保存フォルダー作成",
