+++ resources/i18n/ko.json
@@ -919,6 +919,9 @@
     "CreatedBy": "생성자",
     "Current": "현재",
     "CurrentRevision": "현재 리비전",
+    "CurrentRevisionAvailableDescription": "현재 리비전의 설정을 불러와 새 리비전의 시작점으로 사용하세요.",
+    "CurrentRevisionAvailableTitle": "현재 리비전이 이미 적용 중입니다",
+    "CustomMode": "고급 모드",
     "DeleteDeployment": "배포 삭제",
     "Deploy": "배포",
     "DeployConfirm": "리비전 #{{revisionNumber}}을(를) 배포하시겠습니까? 현재 리비전이 교체됩니다.",
@@ -954,6 +957,7 @@
     "Image": "이미지",
     "LegacyRouteRedirected": "엔드포인트 URL이 변경되었습니다. 새 배포 페이지로 이동했습니다.",
     "LivenessStatus": "활성 상태 (Liveness)",
+    "LoadCurrentRevision": "현재 리비전 불러오기",
     "ManageAccessTokens": "액세스 토큰 관리",
     "Model": "모델",
     "ModelDefinitionPath": "모델 정의 파일 경로",
@@ -969,10 +973,15 @@
     "NewRevisionWillBeCreatedConfirm": "저장하면 새 리비전이 생성됩니다. 계속하시겠습니까?",
     "NoCurrentRevisionDeployed": "배포된 리비전이 없습니다. 리비전을 추가하여 서비스를 활성화하세요.",
     "NoDeployments": "배포가 없습니다.",
+    "NoPresetsAvailable": "사용 가능한 배포 프리셋이 없습니다",
+    "NoPresetsAvailableDescription": "배포에 사용할 프리셋이 없습니다. 대신 새 배포를 생성하세요.",
+    "NoPresetsAvailableSwitchToCustom": "사용 가능한 프리셋이 없습니다. 직접 설정 모드로 전환하여 리비전을 수동으로 추가하세요.",
     "NumberOfReplicas": "레플리카 수",
+    "OpenCreateDeploymentModal": "배포 생성 모달 열기",
     "OpenToPublic": "외부에 공개",
     "Overview": "개요",
     "Owner": "소유자",
+    "PresetMode": "프리셋 모드",
     "Private": "비공개",
     "PrivateDeploymentAlertTitle": "비공개 배포입니다. 엔드포인트를 사용하려면 액세스 토큰을 사용해야 합니다.",
     "Project": "프로젝트",
@@ -1778,6 +1787,7 @@
     "ClearErrors": "오류 지우고 재시도",
     "CloneModelFolderConfirm": "모델 서비스를 실행하려면 폴더를 클론해야 합니다. <strong>\"{{folderName}}\"</strong> 으로 클론 후 서비스를 시작할까요?",
     "Controls": "제어",
+    "CreateNewDeploymentWithPreset": "프리셋으로 새 배포 생성",
     "CreatedAt": "생성 시간",
     "CreatingService": "서비스 생성 중...",
     "CurrentRevisionTitle": "현재 리비전",
@@ -1816,8 +1826,12 @@
     "Image": "이미지",
     "InitialDelay": "초기 지연 시간",
     "InitialDelayTooltip": "서비스 시작 후 첫 번째 건강 체크까지 대기하는 시간(초)입니다.",
+    "Interval": "간격 (s)",
+    "IntervalTooltip": "연속적인 건강 체크 시도 사이의 대기 시간(초)입니다.",
     "MaxRetries": "최대 재시도 횟수",
     "MaxRetriesTooltip": "서비스 실패로 판단하기 전 건강 체크를 재시도하는 최대 횟수입니다.",
+    "MaxWaitTime": "최대 대기 시간 (s)",
+    "MaxWaitTimeTooltip": "단일 건강 체크에 서비스가 응답해야 하는 최대 시간(초)입니다. 이를 초과하면 실패로 처리됩니다.",
     "ModelAndServingConfiguration": "모델 & 서빙 구성",
     "ModelDefinition": "모델 정의",
     "ModelDefinitionPath": "모델 정의 파일 경로",
@@ -1904,7 +1918,6 @@
     "RuntimeVariantNotSpecified": "서비스 정의 파일에 런타임이 지정되지 않았습니다. 계속하려면 배포 프리셋을 선택해 주세요.",
     "RuntimeVariantSelectionRequired": "여러 런타임이 지정되어 있습니다. 계속하려면 배포 프리셋을 선택해 주세요.",
     "SeeContainerLogs": "컨테이너 로그 보기",
-    "SelectDeploymentPreset": "배포 프리셋 선택",
     "ServiceAlreadyExists": "같은 이름의 서비스가 이미 존재합니다",
     "ServiceConfigurationNotValid": "서비스 설정이 올바르지 않습니다. 관리자에게 문의하세요.",
     "ServiceCreated": "모델 서비스 {{ name }}이(가) 생성되었습니다.",
@@ -3022,7 +3035,6 @@
     "Copied": "복사됨",
     "Copy": "복사"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "세션을 만들어 보세요",
     "CreateFolder": "새 스토리지 폴더 생성하기",
