+++ resources/i18n/mn.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Үүсгэсэн хүн",
     "Current": "Одоогийн",
     "CurrentRevision": "Одоогийн Revision",
+    "CurrentRevisionAvailableDescription": "Шинэ Revision-ийн эхлэлийн цэг болгон ашиглахын тулд тохиргоог нь ачаалаарай.",
+    "CurrentRevisionAvailableTitle": "Одоогийн Revision аль хэдийн хүчин төгөлдөр байна",
+    "CustomMode": "Нарийвчилсан горим",
     "DeleteDeployment": "Байршуулалт устгах",
     "Deploy": "Байршуулах",
     "DeployConfirm": "Revision #{{revisionNumber}}-ыг байршуулахдаа итгэлтэй байна уу? Одоогийн Revision солигдох болно.",
@@ -953,6 +956,7 @@
     "Image": "Дүрс",
     "LegacyRouteRedirected": "Endpoint URL шилжсэн байна. Та шинэ Байршуулалт хуудас руу чиглүүлэгдлээ.",
     "LivenessStatus": "Liveness төлөв",
+    "LoadCurrentRevision": "Одоогийн Revision ачаалах",
     "ManageAccessTokens": "Хандалтын токен удирдах",
     "Model": "Загвар",
     "ModelDefinitionPath": "Загварын тодорхойлолтын файлын зам",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Хадгалахад шинэ Revision үүснэ. Үргэлжлүүлэх үү?",
     "NoCurrentRevisionDeployed": "Хэрэгжүүлсэн хувилбар байхгүй — энэ үйлчилгээг идэвхжүүлэхийн тулд хувилбар нэмнэ үү.",
     "NoDeployments": "Байршуулалт олдсонгүй.",
+    "NoPresetsAvailable": "Байршуулалтын урьдчилсан тохиргоо байхгүй",
+    "NoPresetsAvailableDescription": "Байршуулахад ашиглах урьдчилсан тохиргоо байхгүй. Үүний оронд шинэ байршуулалт үүсгэнэ үү.",
+    "NoPresetsAvailableSwitchToCustom": "Боломжтой урьдчилсан тохиргоо байхгүй. Revision-ыг гараар нэмэхийн тулд Тусгай горим руу шилжинэ үү.",
     "NumberOfReplicas": "Репликийн тоо",
+    "OpenCreateDeploymentModal": "Байршуулалт үүсгэх цонхыг нээх",
     "OpenToPublic": "Нийтэд нээлттэй",
     "Overview": "Ерөнхий мэдэгдэл",
     "Owner": "Эзэмшигч",
+    "PresetMode": "Урьдчилсан тохиргооны горим",
     "Private": "Хувийн",
     "PrivateDeploymentAlertTitle": "Хувийн хэрэгжүүлэлт — төгсгөлийн цэгт хандахын тулд хандалтын токен ашиглана уу.",
     "Project": "Төсөл",
@@ -1776,6 +1785,7 @@
     "ClearErrors": "Алдааг арилгаад дахин оролдоно уу",
     "CloneModelFolderConfirm": "Загвар үйлчилгээг ажиллуулахын тулд фолдерыг хувилах хэрэгтэй. \n<strong>\"{{folderName}}\"</strong> -г эхлүүлэх үү?",
     "Controls": "Хяналтууд",
+    "CreateNewDeploymentWithPreset": "Урьдчилан тохируулгаар шинэ байршуулалт үүсгэх",
     "CreatedAt": "Үүсгэсэн",
     "CreatingService": "Үйлчилгээ үүсгэж байна...",
     "CurrentRevisionTitle": "Одоогийн хувилбар",
@@ -1814,8 +1824,12 @@
     "Image": "Зураг",
     "InitialDelay": "Анхны хугацаа",
     "InitialDelayTooltip": "Үйлчилгээ эхэлсний дараа эхний эрүүл мэндийн шалгалтыг хийхийн өмнө хүлээх секунд.",
+    "Interval": "Давтамж (s)",
+    "IntervalTooltip": "Дараалсан эрүүл мэндийн шалгалтуудын хооронд хүлээх секунд.",
     "MaxRetries": "Макс дахин оролдоно",
     "MaxRetriesTooltip": "Үйлчилгээг амжилтгүй гэж үзэхийн өмнөх эрүүл мэндийн шалгалтын дээд оролдлогын тоо.",
+    "MaxWaitTime": "Дээд хүлээлтийн хугацаа (s)",
+    "MaxWaitTimeTooltip": "Нэг эрүүл мэндийн шалгалтад үйлчилгээ хариу өгөхөд зарцуулах дээд секунд. Энэ хугацааг давбал амжилтгүй гэж үзнэ.",
     "ModelAndServingConfiguration": "Загвар & Serving тохиргоо",
     "ModelDefinition": "Загварын тодорхойлолт",
     "ModelDefinitionPath": "Загварын тодорхойлолтын файлын зам",
@@ -1902,7 +1916,6 @@
     "RuntimeVariantNotSpecified": "Үйлчилгээний тодорхойлолтод runtime variant заагдаагүй байна. Үргэлжлүүлэхийн тулд байршуулалтын урьдчилан тохируулгыг сонгоно уу.",
     "RuntimeVariantSelectionRequired": "Олон runtime variant боломжтой байна. Үргэлжлүүлэхийн тулд байршуулалтын урьдчилан тохируулгыг сонгоно уу.",
     "SeeContainerLogs": "Контейнерийн бүртгэлийг үзнэ үү",
-    "SelectDeploymentPreset": "Байршуулалтын урьдчилан тохируулга сонгох",
     "ServiceAlreadyExists": "Үйлчилгээний нэр аль хэдийн байна",
     "ServiceConfigurationNotValid": "Үйлчилгээний тохиргоо буруу байна. \nАдминтай холбоо барина уу.",
     "ServiceCreated": "Загвар үйлчилгээ {{ name }} амжилттай үүсгэгдсэн.",
@@ -3011,7 +3024,6 @@
     "Copied": "Хуулсан",
     "Copy": "Хуулбарлах"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Сеанс үүсгэ",
     "CreateFolder": "Хадгалах шинэ хавтас үүсгэх",
