+++ resources/i18n/ms.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Dicipta Oleh",
     "Current": "Semasa",
     "CurrentRevision": "Revision Semasa",
+    "CurrentRevisionAvailableDescription": "Muatkan konfigurasinya untuk digunakan sebagai titik permulaan Revision baharu.",
+    "CurrentRevisionAvailableTitle": "Revision semasa sudah berkuat kuasa",
+    "CustomMode": "Mod Lanjutan",
     "DeleteDeployment": "Padam Penggunaan",
     "Deploy": "Gunakan",
     "DeployConfirm": "Adakah anda pasti mahu menggunakan Revision #{{revisionNumber}}? Revision semasa akan digantikan.",
@@ -953,6 +956,7 @@
     "Image": "Imej",
     "LegacyRouteRedirected": "URL titik akhir telah berpindah. Anda telah diarahkan ke halaman Penggunaan baharu.",
     "LivenessStatus": "Status Liveness",
+    "LoadCurrentRevision": "Muat Revision semasa",
     "ManageAccessTokens": "Urus token akses",
     "Model": "Model",
     "ModelDefinitionPath": "Laluan Fail Definisi Model",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Menyimpan akan mencipta Revision baharu. Teruskan?",
     "NoCurrentRevisionDeployed": "Tiada semakan yang diterapkan — tambah semakan untuk mengaktifkan perkhidmatan ini.",
     "NoDeployments": "Tiada penggunaan ditemui.",
+    "NoPresetsAvailable": "Tiada praset penggunaan tersedia",
+    "NoPresetsAvailableDescription": "Tiada praset untuk menggunakan dengan. Cipta penggunaan baharu sebaliknya.",
+    "NoPresetsAvailableSwitchToCustom": "Tiada praset tersedia. Tukar ke mod Tersuai untuk menambah Revision secara manual.",
     "NumberOfReplicas": "Bilangan Replika",
+    "OpenCreateDeploymentModal": "Buka modal cipta penggunaan",
     "OpenToPublic": "Terbuka kepada Awam",
     "Overview": "Gambaran Keseluruhan",
     "Owner": "Pemilik",
+    "PresetMode": "Mod Praset",
     "Private": "Peribadi",
     "PrivateDeploymentAlertTitle": "Penerapan persendirian — gunakan token akses untuk mengakses titik akhir.",
     "Project": "Projek",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Kosongkan ralat dan cuba semula",
     "CloneModelFolderConfirm": "Untuk menjalankan perkhidmatan model, folder perlu diklon. \nAdakah anda ingin mengklonkannya sebagai <strong>\"{{folderName}}\"</strong> dan memulakan perkhidmatan?",
     "Controls": "Kawalan",
+    "CreateNewDeploymentWithPreset": "Cipta Penggunaan Baharu dengan Praset",
     "CreatedAt": "Dicipta Pada",
     "CreatingService": "Mencipta perkhidmatan...",
     "CurrentRevisionTitle": "Semakan Semasa",
@@ -1813,8 +1823,12 @@
     "Image": "Gambar",
     "InitialDelay": "Kelewatan Awal",
     "InitialDelayTooltip": "Saat untuk menunggu sebelum semakan kesihatan pertama selepas perkhidmatan dimulakan.",
+    "Interval": "Selang (s)",
+    "IntervalTooltip": "Saat antara percubaan semakan kesihatan yang berturutan.",
     "MaxRetries": "Max Cuba Semula",
     "MaxRetriesTooltip": "Bilangan maksimum percubaan semakan kesihatan sebelum perkhidmatan dianggap gagal.",
+    "MaxWaitTime": "Masa Tunggu Maks (s)",
+    "MaxWaitTimeTooltip": "Saat maksimum yang dimiliki perkhidmatan untuk bertindak balas terhadap semakan kesihatan tunggal sebelum dianggap sebagai kegagalan.",
     "ModelAndServingConfiguration": "Konfigurasi Model & Serving",
     "ModelDefinition": "Definisi Model",
     "ModelDefinitionPath": "Laluan Fail Definisi Model",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "Varian runtime tidak dinyatakan dalam definisi perkhidmatan. Sila pilih pratetap penggunaan untuk meneruskan.",
     "RuntimeVariantSelectionRequired": "Beberapa varian runtime tersedia. Sila pilih pratetap penggunaan untuk meneruskan.",
     "SeeContainerLogs": "Lihat Log Bekas",
-    "SelectDeploymentPreset": "Pilih Pratetap Penggunaan",
     "ServiceAlreadyExists": "Nama perkhidmatan sudah ada",
     "ServiceConfigurationNotValid": "Konfigurasi perkhidmatan tidak sah. \nSila hubungi pentadbir.",
     "ServiceCreated": "Perkhidmatan model {{ name }} telah berjaya dibuat.",
@@ -3009,7 +3022,6 @@
     "Copied": "disalin",
     "Copy": "Salinan"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Cipta Sesi",
     "CreateFolder": "Buat Folder Storan Baharu",
