+++ resources/i18n/pl.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Utworzone przez",
     "Current": "Bieżąca",
     "CurrentRevision": "Bieżąca Revision",
+    "CurrentRevisionAvailableDescription": "Załaduj jej konfigurację, aby użyć jej jako punktu startowego nowej Revision.",
+    "CurrentRevisionAvailableTitle": "Bieżąca Revision jest już aktywna",
+    "CustomMode": "Tryb zaawansowany",
     "DeleteDeployment": "Usuń wdrożenie",
     "Deploy": "Wdróż",
     "DeployConfirm": "Czy na pewno chcesz wdrożyć Revision #{{revisionNumber}}? Bieżąca Revision zostanie zastąpiona.",
@@ -953,6 +956,7 @@
     "Image": "Obraz",
     "LegacyRouteRedirected": "URL punktu końcowego został przeniesiony. Zostałeś przekierowany na nową stronę Wdrożeń.",
     "LivenessStatus": "Status Liveness",
+    "LoadCurrentRevision": "Załaduj bieżącą Revision",
     "ManageAccessTokens": "Zarządzaj tokenami dostępu",
     "Model": "Model",
     "ModelDefinitionPath": "Ścieżka pliku definicji modelu",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Zapisanie spowoduje utworzenie nowej Revision. Kontynuować?",
     "NoCurrentRevisionDeployed": "Nie wdrożono żadnej rewizji — dodaj rewizję, aby aktywować tę usługę.",
     "NoDeployments": "Nie znaleziono wdrożeń.",
+    "NoPresetsAvailable": "Brak dostępnych presetów wdrożenia",
+    "NoPresetsAvailableDescription": "Nie ma presetu do wdrożenia. Zamiast tego utwórz nowe wdrożenie.",
+    "NoPresetsAvailableSwitchToCustom": "Brak dostępnych presetów. Przełącz się na tryb niestandardowy, aby dodać Revision ręcznie.",
     "NumberOfReplicas": "Liczba replik",
+    "OpenCreateDeploymentModal": "Otwórz modal tworzenia wdrożenia",
     "OpenToPublic": "Publiczne",
     "Overview": "Przegląd",
     "Owner": "Właściciel",
+    "PresetMode": "Tryb presetu",
     "Private": "Prywatne",
     "PrivateDeploymentAlertTitle": "Wdrożenie prywatne — użyj tokenu dostępu, aby uzyskać dostęp do punktu końcowego.",
     "Project": "Projekt",
@@ -1776,6 +1785,7 @@
     "ClearErrors": "Usuń błąd i ponów próbę",
     "CloneModelFolderConfirm": "Aby uruchomić usługę modelową, folder musi zostać sklonowany. \nCzy chciałbyś to klonizować jako <strong>\"{{folderName}}\"</strong> i uruchomić usługę?",
     "Controls": "Elementy sterujące",
+    "CreateNewDeploymentWithPreset": "Utwórz nowe wdrożenie z presetem",
     "CreatedAt": "Utworzono w",
     "CreatingService": "Tworzenie usługi...",
     "CurrentRevisionTitle": "Bieżąca rewizja",
@@ -1814,8 +1824,12 @@
     "Image": "Obraz",
     "InitialDelay": "Opóźnienie początkowe",
     "InitialDelayTooltip": "Sekundy oczekiwania przed pierwszą kontrolą zdrowia po uruchomieniu usługi.",
+    "Interval": "Interwał (s)",
+    "IntervalTooltip": "Sekundy między kolejnymi próbami kontroli zdrowia.",
     "MaxRetries": "Maksymalna liczba ponownych prób",
     "MaxRetriesTooltip": "Maksymalna liczba prób kontroli zdrowia przed uznaniem usługi za niedziałającą.",
+    "MaxWaitTime": "Maksymalny czas oczekiwania (s)",
+    "MaxWaitTimeTooltip": "Maksymalna liczba sekund, jaką usługa ma na odpowiedź na pojedynczą kontrolę zdrowia, zanim zostanie uznana za nieudaną.",
     "ModelAndServingConfiguration": "Konfiguracja modelu i serwowania",
     "ModelDefinition": "Definicja modelu",
     "ModelDefinitionPath": "Ścieżka pliku definicji modelu",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "Wariant runtime nie jest określony w definicji usługi. Wybierz preset wdrożenia, aby kontynuować.",
     "RuntimeVariantSelectionRequired": "Dostępnych jest wiele wariantów runtime. Wybierz preset wdrożenia, aby kontynuować.",
     "SeeContainerLogs": "Zobacz dzienniki kontenerów",
-    "SelectDeploymentPreset": "Wybierz preset wdrożenia",
     "ServiceAlreadyExists": "Nazwa usługi już istnieje",
     "ServiceConfigurationNotValid": "Konfiguracja usługi jest nieprawidłowa. \nSkontaktuj się z administratorem.",
     "ServiceCreated": "Usługa modelowa {{ name }} została pomyślnie utworzona.",
@@ -3012,7 +3025,6 @@
     "Copied": "Skopiowano",
     "Copy": "Kopiuj"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Utwórz sesję",
     "CreateFolder": "Utwórz nowy folder przechowywania",
