+++ resources/i18n/pt-BR.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Criado por",
     "Current": "Atual",
     "CurrentRevision": "Revision Atual",
+    "CurrentRevisionAvailableDescription": "Carregue sua configuração para usá-la como ponto de partida da nova Revision.",
+    "CurrentRevisionAvailableTitle": "Uma Revision atual já está em vigor",
+    "CustomMode": "Modo Avançado",
     "DeleteDeployment": "Excluir implantação",
     "Deploy": "Implantar",
     "DeployConfirm": "Tem certeza de que deseja implantar a Revision #{{revisionNumber}}? A Revision atual será substituída.",
@@ -953,6 +956,7 @@
     "Image": "Imagem",
     "LegacyRouteRedirected": "A URL do endpoint foi movida. Você foi redirecionado para a nova página de Implantações.",
     "LivenessStatus": "Status de Liveness",
+    "LoadCurrentRevision": "Carregar Revision atual",
     "ManageAccessTokens": "Gerenciar tokens de acesso",
     "Model": "Modelo",
     "ModelDefinitionPath": "Caminho do arquivo de definição do modelo",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Salvar criará uma nova Revision. Continuar?",
     "NoCurrentRevisionDeployed": "Nenhuma revisão implantada — adicione uma revisão para ativar este serviço.",
     "NoDeployments": "Nenhuma implantação encontrada.",
+    "NoPresetsAvailable": "Nenhum preset de implantação disponível",
+    "NoPresetsAvailableDescription": "Não há nenhum preset para implantar. Crie uma nova implantação.",
+    "NoPresetsAvailableSwitchToCustom": "Nenhum preset disponível. Mude para o modo Personalizado para adicionar uma Revision manualmente.",
     "NumberOfReplicas": "Número de réplicas",
+    "OpenCreateDeploymentModal": "Abrir modal de criação de implantação",
     "OpenToPublic": "Público",
     "Overview": "Visão geral",
     "Owner": "Proprietário",
+    "PresetMode": "Modo Preset",
     "Private": "Privado",
     "PrivateDeploymentAlertTitle": "Implantação privada — use um token de acesso para acessar o endpoint.",
     "Project": "Projeto",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Limpar o erro e tentar de novo",
     "CloneModelFolderConfirm": "Para executar o serviço de modelo, a pasta precisa ser clonada. \nVocê gostaria de cloná -lo como <strong>\"{{folderName}}\"</strong> e iniciar o serviço?",
     "Controls": "Controlos",
+    "CreateNewDeploymentWithPreset": "Criar nova implantação com preset",
     "CreatedAt": "Criado em",
     "CreatingService": "Criando serviço...",
     "CurrentRevisionTitle": "Revisão atual",
@@ -1813,8 +1823,12 @@
     "Image": "Imagem",
     "InitialDelay": "Atraso inicial",
     "InitialDelayTooltip": "Segundos para aguardar antes da primeira verificação de integridade após o início do serviço.",
+    "Interval": "Intervalo (s)",
+    "IntervalTooltip": "Segundos entre tentativas consecutivas de verificação de integridade.",
     "MaxRetries": "Máximo de tentativas",
     "MaxRetriesTooltip": "Número máximo de tentativas de verificação de integridade antes de o serviço ser considerado falho.",
+    "MaxWaitTime": "Tempo máx. de espera (s)",
+    "MaxWaitTimeTooltip": "Segundos máximos que o serviço tem para responder a uma única verificação de integridade antes de ser tratado como falha.",
     "ModelAndServingConfiguration": "Configuração de modelo e serviço",
     "ModelDefinition": "Definição do modelo",
     "ModelDefinitionPath": "Caminho do arquivo de definição de modelo",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "A variante de runtime não está especificada na definição do serviço. Selecione um preset de implantação para continuar.",
     "RuntimeVariantSelectionRequired": "Várias variantes de runtime estão disponíveis. Selecione um preset de implantação para continuar.",
     "SeeContainerLogs": "Consulte os registros do contêiner",
-    "SelectDeploymentPreset": "Selecionar preset de implantação",
     "ServiceAlreadyExists": "O nome do serviço já existe",
     "ServiceConfigurationNotValid": "A configuração do serviço é inválida. \nEntre em contato com o administrador.",
     "ServiceCreated": "O serviço de modelo {{ name }} foi criado com sucesso.",
@@ -3009,7 +3022,6 @@
     "Copied": "Copiado",
     "Copy": "cópia de"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Criar uma sessão",
     "CreateFolder": "Crie uma nova pasta de armazenamento",
