+++ resources/i18n/pt.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Criado por",
     "Current": "Atual",
     "CurrentRevision": "Revision Atual",
+    "CurrentRevisionAvailableDescription": "Carregue a sua configuração para a utilizar como ponto de partida para a nova Revision.",
+    "CurrentRevisionAvailableTitle": "Uma Revision atual já está em vigor",
+    "CustomMode": "Modo Avançado",
     "DeleteDeployment": "Eliminar implementação",
     "Deploy": "Implementar",
     "DeployConfirm": "Tem a certeza de que pretende implementar a Revision #{{revisionNumber}}? A Revision atual será substituída.",
@@ -953,6 +956,7 @@
     "Image": "Imagem",
     "LegacyRouteRedirected": "O URL do endpoint foi movido. Foi redirecionado para a nova página de Implementações.",
     "LivenessStatus": "Estado de Liveness",
+    "LoadCurrentRevision": "Carregar Revision atual",
     "ManageAccessTokens": "Gerir tokens de acesso",
     "Model": "Modelo",
     "ModelDefinitionPath": "Caminho do ficheiro de definição do modelo",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Guardar criará uma nova Revision. Continuar?",
     "NoCurrentRevisionDeployed": "Nenhuma revisão implementada — adicione uma revisão para ativar este serviço.",
     "NoDeployments": "Nenhuma implementação encontrada.",
+    "NoPresetsAvailable": "Nenhum preset de implementação disponível",
+    "NoPresetsAvailableDescription": "Não existe nenhum preset para implementar. Crie uma nova implementação.",
+    "NoPresetsAvailableSwitchToCustom": "Nenhum preset disponível. Mude para o modo Personalizado para adicionar uma Revision manualmente.",
     "NumberOfReplicas": "Número de réplicas",
+    "OpenCreateDeploymentModal": "Abrir modal de criação de implementação",
     "OpenToPublic": "Público",
     "Overview": "Visão geral",
     "Owner": "Proprietário",
+    "PresetMode": "Modo Preset",
     "Private": "Privado",
     "PrivateDeploymentAlertTitle": "Implementação privada — utilize um token de acesso para aceder ao endpoint.",
     "Project": "Projeto",
@@ -1777,6 +1786,7 @@
     "ClearErrors": "Limpar o erro e tentar de novo",
     "CloneModelFolderConfirm": "Para executar o serviço de modelo, a pasta precisa ser clonada. \nVocê gostaria de cloná -lo como <strong>\"{{folderName}}\"</strong> e iniciar o serviço?",
     "Controls": "Controlos",
+    "CreateNewDeploymentWithPreset": "Criar nova implementação com preset",
     "CreatedAt": "Criado em",
     "CreatingService": "Creating service...",
     "CurrentRevisionTitle": "Revisão atual",
@@ -1815,8 +1825,12 @@
     "Image": "Imagem",
     "InitialDelay": "Atraso inicial",
     "InitialDelayTooltip": "Segundos para aguardar antes da primeira verificação de integridade após o início do serviço.",
+    "Interval": "Intervalo (s)",
+    "IntervalTooltip": "Segundos entre tentativas consecutivas de verificação de integridade.",
     "MaxRetries": "Máximo de tentativas",
     "MaxRetriesTooltip": "Número máximo de tentativas de verificação de integridade antes de o serviço ser considerado com falha.",
+    "MaxWaitTime": "Tempo máx. de espera (s)",
+    "MaxWaitTimeTooltip": "Segundos máximos que o serviço tem para responder a uma única verificação de integridade antes de ser tratado como falha.",
     "ModelAndServingConfiguration": "Configuração de modelo e serviço",
     "ModelDefinition": "Definição do modelo",
     "ModelDefinitionPath": "Caminho do arquivo de definição de modelo",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "A variante de runtime não está especificada na definição de serviço. Selecione um preset de implementação para continuar.",
     "RuntimeVariantSelectionRequired": "Estão disponíveis múltiplas variantes de runtime. Selecione um preset de implementação para continuar.",
     "SeeContainerLogs": "Consulte os registros do contêiner",
-    "SelectDeploymentPreset": "Selecionar preset de implementação",
     "ServiceAlreadyExists": "O nome do serviço já existe",
     "ServiceConfigurationNotValid": "A configuração do serviço é inválida. \nEntre em contato com o administrador.",
     "ServiceCreated": "O serviço de modelo {{ name }} foi criado com sucesso.",
@@ -3011,7 +3024,6 @@
     "Copied": "Copiado",
     "Copy": "cópia de"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Criar uma sessão",
     "CreateFolder": "Crie uma nova pasta de armazenamento",
