+++ resources/i18n/ru.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Создано пользователем",
     "Current": "Текущая",
     "CurrentRevision": "Текущая Revision",
+    "CurrentRevisionAvailableDescription": "Загрузите её конфигурацию, чтобы использовать её как отправную точку для новой Revision.",
+    "CurrentRevisionAvailableTitle": "Текущая Revision уже действует",
+    "CustomMode": "Расширенный режим",
     "DeleteDeployment": "Удалить развёртывание",
     "Deploy": "Развернуть",
     "DeployConfirm": "Вы уверены, что хотите развернуть Revision #{{revisionNumber}}? Текущая Revision будет заменена.",
@@ -953,6 +956,7 @@
     "Image": "Образ",
     "LegacyRouteRedirected": "URL конечной точки изменён. Вы перенаправлены на новую страницу развёртываний.",
     "LivenessStatus": "Статус Liveness",
+    "LoadCurrentRevision": "Загрузить текущую Revision",
     "ManageAccessTokens": "Управление токенами доступа",
     "Model": "Модель",
     "ModelDefinitionPath": "Путь к файлу определения модели",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Сохранение создаст новую Revision. Продолжить?",
     "NoCurrentRevisionDeployed": "Ревизия не развернута — добавьте ревизию, чтобы активировать этот сервис.",
     "NoDeployments": "Развёртывания не найдены.",
+    "NoPresetsAvailable": "Нет доступных пресетов развёртывания",
+    "NoPresetsAvailableDescription": "Нет пресетов для развёртывания. Создайте вместо этого новое развёртывание.",
+    "NoPresetsAvailableSwitchToCustom": "Нет доступных пресетов. Переключитесь в пользовательский режим, чтобы добавить Revision вручную.",
     "NumberOfReplicas": "Количество реплик",
+    "OpenCreateDeploymentModal": "Открыть модальное окно создания развёртывания",
     "OpenToPublic": "Публичное",
     "Overview": "Обзор",
     "Owner": "Владелец",
+    "PresetMode": "Режим пресета",
     "Private": "Приватный",
     "PrivateDeploymentAlertTitle": "Приватное развертывание — используйте токен доступа для обращения к конечной точке.",
     "Project": "Проект",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Очистить ошибку и повторить попытку",
     "CloneModelFolderConfirm": "Чтобы запустить услугу модели, папку нужно клонировать. \nХотели бы вы клонировать его как <strong>\"{{folderName}}\"</strong> и запустить службу?",
     "Controls": "Средства управления",
+    "CreateNewDeploymentWithPreset": "Создать новое развёртывание с пресетом",
     "CreatedAt": "Создано в",
     "CreatingService": "Создание сервиса...",
     "CurrentRevisionTitle": "Текущая ревизия",
@@ -1813,8 +1823,12 @@
     "Image": "Изображение",
     "InitialDelay": "Начальная задержка",
     "InitialDelayTooltip": "Количество секунд ожидания перед первой проверкой работоспособности после запуска сервиса.",
+    "Interval": "Интервал (с)",
+    "IntervalTooltip": "Секунды между последовательными попытками проверки работоспособности.",
     "MaxRetries": "Макс. повторов",
     "MaxRetriesTooltip": "Максимальное количество попыток проверки работоспособности до признания сервиса неработающим.",
+    "MaxWaitTime": "Макс. время ожидания (с)",
+    "MaxWaitTimeTooltip": "Максимальное количество секунд, отведённых сервису для ответа на одну проверку работоспособности, прежде чем она будет считаться неудачной.",
     "ModelAndServingConfiguration": "Конфигурация модели и обслуживания",
     "ModelDefinition": "Определение модели",
     "ModelDefinitionPath": "Путь к файлу определения модели",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "Вариант runtime не указан в определении сервиса. Для продолжения выберите предустановку развёртывания.",
     "RuntimeVariantSelectionRequired": "Доступно несколько вариантов runtime. Для продолжения выберите предустановку развёртывания.",
     "SeeContainerLogs": "См. журналы контейнеров",
-    "SelectDeploymentPreset": "Выбрать предустановку развёртывания",
     "ServiceAlreadyExists": "Название услуги уже существует",
     "ServiceConfigurationNotValid": "Конфигурация службы недействительна. \nПожалуйста, свяжитесь с администратором.",
     "ServiceCreated": "Сервис модели {{ name }} успешно создан.",
@@ -3009,7 +3022,6 @@
     "Copied": "Скопировано",
     "Copy": "Копировать"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Создать сеанс",
     "CreateFolder": "Создайте новую папку хранения",
