+++ resources/i18n/th.json
@@ -918,6 +918,9 @@
     "CreatedBy": "สร้างโดย",
     "Current": "ปัจจุบัน",
     "CurrentRevision": "Revision ปัจจุบัน",
+    "CurrentRevisionAvailableDescription": "โหลดการกำหนดค่าเพื่อใช้เป็นจุดเริ่มต้นสำหรับ Revision ใหม่",
+    "CurrentRevisionAvailableTitle": "มี Revision ปัจจุบันที่มีผลอยู่แล้ว",
+    "CustomMode": "โหมดขั้นสูง",
     "DeleteDeployment": "ลบการปรับใช้",
     "Deploy": "ปรับใช้",
     "DeployConfirm": "คุณแน่ใจหรือไม่ว่าต้องการปรับใช้ Revision #{{revisionNumber}}? Revision ปัจจุบันจะถูกแทนที่",
@@ -953,6 +956,7 @@
     "Image": "อิมเมจ",
     "LegacyRouteRedirected": "URL ปลายทางถูกย้ายแล้ว คุณถูกนำไปยังหน้าการปรับใช้ใหม่",
     "LivenessStatus": "สถานะ Liveness",
+    "LoadCurrentRevision": "โหลด Revision ปัจจุบัน",
     "ManageAccessTokens": "จัดการโทเค็นการเข้าถึง",
     "Model": "โมเดล",
     "ModelDefinitionPath": "เส้นทางไฟล์นิยามโมเดล",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "การบันทึกจะสร้าง Revision ใหม่ ดำเนินการต่อหรือไม่?",
     "NoCurrentRevisionDeployed": "ไม่มีการปรับใช้รุ่นแก้ไข — เพิ่มรุ่นแก้ไขเพื่อเปิดใช้งานบริการนี้",
     "NoDeployments": "ไม่พบการปรับใช้",
+    "NoPresetsAvailable": "ไม่มีพรีเซตการปรับใช้",
+    "NoPresetsAvailableDescription": "ไม่มีพรีเซตสำหรับการปรับใช้ สร้างการปรับใช้ใหม่แทน",
+    "NoPresetsAvailableSwitchToCustom": "ไม่มีพรีเซตที่พร้อมใช้งาน สลับไปยังโหมดกำหนดเองเพื่อเพิ่ม Revision ด้วยตนเอง",
     "NumberOfReplicas": "จำนวนเรพลิกา",
+    "OpenCreateDeploymentModal": "เปิดกล่องโต้ตอบสร้างการปรับใช้",
     "OpenToPublic": "เปิดสาธารณะ",
     "Overview": "ภาพรวม",
     "Owner": "เจ้าของ",
+    "PresetMode": "โหมดพรีเซต",
     "Private": "ส่วนตัว",
     "PrivateDeploymentAlertTitle": "การปรับใช้แบบส่วนตัว — ใช้โทเค็นการเข้าถึงเพื่อเข้าถึงปลายทาง",
     "Project": "โปรเจกต์",
@@ -1777,6 +1786,7 @@
     "ClearErrors": "ล้างข้อผิดพลาดและลองใหม่",
     "CloneModelFolderConfirm": "ในการเรียกใช้บริการรุ่นโฟลเดอร์จะต้องถูกโคลน \nคุณต้องการโคลนเป็น <strong>\"{{folderName}}\"</strong> และเริ่มบริการหรือไม่?",
     "Controls": "การควบคุม",
+    "CreateNewDeploymentWithPreset": "สร้าง Deployment ใหม่ด้วย Preset",
     "CreatedAt": "สร้างเมื่อ",
     "CreatingService": "Creating service...",
     "CurrentRevisionTitle": "รีวิชันปัจจุบัน",
@@ -1815,8 +1825,12 @@
     "Image": "อิมเมจ",
     "InitialDelay": "ความล่าช้าเริ่มต้น",
     "InitialDelayTooltip": "จำนวนวินาทีที่รอก่อนการตรวจสอบสถานะครั้งแรกหลังจากบริการเริ่มทำงาน",
+    "Interval": "ช่วงเวลา (s)",
+    "IntervalTooltip": "จำนวนวินาทีระหว่างการพยายามตรวจสอบสถานะแต่ละครั้ง",
     "MaxRetries": "จำนวนครั้งที่ลองใหม่สูงสุด",
     "MaxRetriesTooltip": "จำนวนครั้งสูงสุดในการพยายามตรวจสอบสถานะก่อนที่บริการจะถูกถือว่าล้มเหลว",
+    "MaxWaitTime": "เวลารอสูงสุด (s)",
+    "MaxWaitTimeTooltip": "จำนวนวินาทีสูงสุดที่บริการมีในการตอบสนองต่อการตรวจสอบสถานะครั้งเดียวก่อนที่จะถือว่าล้มเหลว",
     "ModelAndServingConfiguration": "การตั้งค่าโมเดลและการบริการ",
     "ModelDefinition": "คำนิยามโมเดล",
     "ModelDefinitionPath": "เส้นทางไฟล์กำหนดโมเดล",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "ไม่ได้ระบุ runtime variant ในนิยามบริการ กรุณาเลือก deployment preset เพื่อดำเนินการต่อ",
     "RuntimeVariantSelectionRequired": "มี runtime variant หลายตัวให้เลือก กรุณาเลือก deployment preset เพื่อดำเนินการต่อ",
     "SeeContainerLogs": "ดูบันทึกคอนเทนเนอร์",
-    "SelectDeploymentPreset": "เลือก Deployment Preset",
     "ServiceAlreadyExists": "ชื่อบริการมีอยู่แล้ว",
     "ServiceConfigurationNotValid": "การกำหนดค่าบริการไม่ถูกต้อง \nกรุณาติดต่อผู้ดูแลระบบ",
     "ServiceCreated": "สร้างบริการโมเดล {{ name }} สำเร็จแล้ว",
@@ -3011,7 +3024,6 @@
     "Copied": "คัดลอกแล้ว",
     "Copy": "คัดลอก"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "สร้างเซสชัน",
     "CreateFolder": "สร้างโฟลเดอร์จัดเก็บข้อมูลใหม่",
