+++ resources/i18n/tr.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Oluşturan",
     "Current": "Mevcut",
     "CurrentRevision": "Mevcut Revision",
+    "CurrentRevisionAvailableDescription": "Yeni Revision için başlangıç noktası olarak kullanmak üzere konfigürasyonunu yükleyin.",
+    "CurrentRevisionAvailableTitle": "Mevcut bir Revision zaten yürürlükte",
+    "CustomMode": "Gelişmiş Mod",
     "DeleteDeployment": "Dağıtımı Sil",
     "Deploy": "Dağıt",
     "DeployConfirm": "Revision #{{revisionNumber}}'ı dağıtmak istediğinizden emin misiniz? Mevcut Revision değiştirilecek.",
@@ -953,6 +956,7 @@
     "Image": "İmaj",
     "LegacyRouteRedirected": "Uç nokta URL'si taşındı. Yeni Dağıtımlar sayfasına yönlendirildiniz.",
     "LivenessStatus": "Liveness Durumu",
+    "LoadCurrentRevision": "Mevcut Revision'ı yükle",
     "ManageAccessTokens": "Erişim belirteçlerini yönet",
     "Model": "Model",
     "ModelDefinitionPath": "Model Tanım Dosyası Yolu",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Kaydetme yeni bir Revision oluşturacak. Devam edilsin mi?",
     "NoCurrentRevisionDeployed": "Dağıtılmış revizyon yok — bu hizmeti etkinleştirmek için bir revizyon ekleyin.",
     "NoDeployments": "Dağıtım bulunamadı.",
+    "NoPresetsAvailable": "Kullanılabilir dağıtım ön ayarı yok",
+    "NoPresetsAvailableDescription": "Dağıtım için kullanılabilecek ön ayar bulunmuyor. Bunun yerine yeni bir dağıtım oluşturun.",
+    "NoPresetsAvailableSwitchToCustom": "Kullanılabilir ön ayar yok. Revision'ı manuel olarak eklemek için Özel moda geçin.",
     "NumberOfReplicas": "Replika Sayısı",
+    "OpenCreateDeploymentModal": "Dağıtım oluşturma modalını aç",
     "OpenToPublic": "Herkese Açık",
     "Overview": "Genel Bakış",
     "Owner": "Sahibi",
+    "PresetMode": "Ön Ayar Modu",
     "Private": "Özel",
     "PrivateDeploymentAlertTitle": "Özel dağıtım — uç noktasına erişmek için bir erişim belirteci kullanın.",
     "Project": "Proje",
@@ -1775,6 +1784,7 @@
     "ClearErrors": "Hatayı temizle ve yeniden dene",
     "CloneModelFolderConfirm": "Model hizmetini çalıştırmak için klasörün klonlanması gerekir. \nBunu <strong>\"{{folderName}}\"</strong> olarak klonlamak ve hizmeti başlatmak ister misiniz?",
     "Controls": "Kontroller",
+    "CreateNewDeploymentWithPreset": "Ön Ayar ile Yeni Dağıtım Oluştur",
     "CreatedAt": "Şu Adreste Oluşturuldu",
     "CreatingService": "Hizmet oluşturuluyor...",
     "CurrentRevisionTitle": "Mevcut Revizyon",
@@ -1813,8 +1823,12 @@
     "Image": "Resim",
     "InitialDelay": "Başlangıç gecikmesi",
     "InitialDelayTooltip": "Hizmet başladıktan sonra ilk sağlık kontrolünden önce beklenecek saniye sayısı.",
+    "Interval": "Aralık (s)",
+    "IntervalTooltip": "Ardışık sağlık kontrolü denemeleri arasındaki saniye sayısı.",
     "MaxRetries": "Maksimum Yeniden Deneme Sayısı",
     "MaxRetriesTooltip": "Hizmet başarısız sayılmadan önce gerçekleştirilebilecek maksimum sağlık kontrolü deneme sayısı.",
+    "MaxWaitTime": "Maksimum Bekleme Süresi (s)",
+    "MaxWaitTimeTooltip": "Hizmetin tek bir sağlık kontrolüne yanıt vermesi için ayrılan maksimum saniye sayısı. Bu süre aşılırsa başarısız kabul edilir.",
     "ModelAndServingConfiguration": "Model & Serving Yapılandırması",
     "ModelDefinition": "Model tanımı",
     "ModelDefinitionPath": "Model Tanımı Dosya Yolu",
@@ -1901,7 +1915,6 @@
     "RuntimeVariantNotSpecified": "Servis tanımında runtime varyantı belirtilmemiş. Devam etmek için bir dağıtım ön ayarı seçin.",
     "RuntimeVariantSelectionRequired": "Birden fazla runtime varyantı mevcut. Devam etmek için bir dağıtım ön ayarı seçin.",
     "SeeContainerLogs": "Konteyner Günlüklerini görün",
-    "SelectDeploymentPreset": "Dağıtım Ön Ayarı Seçin",
     "ServiceAlreadyExists": "Hizmet Adı zaten var",
     "ServiceConfigurationNotValid": "Hizmet yapılandırması geçersiz. \nLütfen yöneticiyle iletişime geçin.",
     "ServiceCreated": "Model hizmeti {{ name }} başarıyla oluşturuldu.",
@@ -3009,7 +3022,6 @@
     "Copied": "Kopyalandı",
     "Copy": "Kopyala"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Oturum Oluşturma",
     "CreateFolder": "Yeni Bir Depolama Klasörü Oluşturun",
