+++ resources/i18n/vi.json
@@ -918,6 +918,9 @@
     "CreatedBy": "Tạo bởi",
     "Current": "Hiện tại",
     "CurrentRevision": "Revision Hiện Tại",
+    "CurrentRevisionAvailableDescription": "Tải cấu hình của nó để sử dụng làm điểm khởi đầu cho Revision mới.",
+    "CurrentRevisionAvailableTitle": "Đã có một Revision hiện tại đang có hiệu lực",
+    "CustomMode": "Chế độ nâng cao",
     "DeleteDeployment": "Xóa triển khai",
     "Deploy": "Triển khai",
     "DeployConfirm": "Bạn có chắc chắn muốn triển khai Revision #{{revisionNumber}} không? Revision hiện tại sẽ được thay thế.",
@@ -953,6 +956,7 @@
     "Image": "Hình ảnh",
     "LegacyRouteRedirected": "URL điểm cuối đã được di chuyển. Bạn đã được chuyển hướng đến trang Triển khai mới.",
     "LivenessStatus": "Trạng thái Liveness",
+    "LoadCurrentRevision": "Tải Revision hiện tại",
     "ManageAccessTokens": "Quản lý mã thông báo truy cập",
     "Model": "Mô hình",
     "ModelDefinitionPath": "Đường dẫn tệp định nghĩa mô hình",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "Lưu sẽ tạo Revision mới. Tiếp tục?",
     "NoCurrentRevisionDeployed": "Không có bản sửa đổi nào được triển khai — thêm bản sửa đổi để kích hoạt dịch vụ này.",
     "NoDeployments": "Không tìm thấy triển khai.",
+    "NoPresetsAvailable": "Không có preset triển khai nào",
+    "NoPresetsAvailableDescription": "Không có preset nào để triển khai. Hãy tạo một triển khai mới.",
+    "NoPresetsAvailableSwitchToCustom": "Không có preset nào. Hãy chuyển sang chế độ Tùy chỉnh để thêm Revision thủ công.",
     "NumberOfReplicas": "Số lượng bản sao",
+    "OpenCreateDeploymentModal": "Mở hộp thoại tạo triển khai",
     "OpenToPublic": "Công khai",
     "Overview": "Tổng quan",
     "Owner": "Chủ sở hữu",
+    "PresetMode": "Chế độ Preset",
     "Private": "Riêng tư",
     "PrivateDeploymentAlertTitle": "Triển khai riêng tư — sử dụng mã thông báo truy cập để truy cập điểm cuối.",
     "Project": "Dự án",
@@ -1777,6 +1786,7 @@
     "ClearErrors": "Xóa lỗi và thử lại",
     "CloneModelFolderConfirm": "Để chạy dịch vụ mô hình, thư mục cần được nhân bản. \nBạn có muốn sao chép nó dưới dạng <strong>\"{{folderName}}\"</strong> và bắt đầu dịch vụ không?",
     "Controls": "Điều khiển",
+    "CreateNewDeploymentWithPreset": "Tạo triển khai mới với Preset",
     "CreatedAt": "Được tạo tại",
     "CreatingService": "Đang tạo dịch vụ...",
     "CurrentRevisionTitle": "Phiên bản hiện tại",
@@ -1815,8 +1825,12 @@
     "Image": "Hình ảnh",
     "InitialDelay": "Độ trễ ban đầu",
     "InitialDelayTooltip": "Số giây chờ trước lần kiểm tra sức khỏe đầu tiên sau khi dịch vụ khởi động.",
+    "Interval": "Khoảng thời gian (s)",
+    "IntervalTooltip": "Số giây giữa các lần thử kiểm tra sức khỏe liên tiếp.",
     "MaxRetries": "Số lần thử lại tối đa",
     "MaxRetriesTooltip": "Số lần thử kiểm tra sức khỏe tối đa trước khi dịch vụ được coi là thất bại.",
+    "MaxWaitTime": "Thời gian chờ tối đa (s)",
+    "MaxWaitTimeTooltip": "Số giây tối đa để dịch vụ phản hồi một lần kiểm tra sức khỏe trước khi bị coi là thất bại.",
     "ModelAndServingConfiguration": "Cấu hình mô hình & phục vụ",
     "ModelDefinition": "Định nghĩa mô hình",
     "ModelDefinitionPath": "Đường dẫn tệp định nghĩa mô hình",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "Biến thể runtime không được chỉ định trong định nghĩa dịch vụ. Vui lòng chọn preset triển khai để tiếp tục.",
     "RuntimeVariantSelectionRequired": "Có nhiều biến thể runtime khả dụng. Vui lòng chọn preset triển khai để tiếp tục.",
     "SeeContainerLogs": "Xem nhật ký vùng chứa",
-    "SelectDeploymentPreset": "Chọn Preset Triển Khai",
     "ServiceAlreadyExists": "Tên dịch vụ đã tồn tại",
     "ServiceConfigurationNotValid": "Cấu hình dịch vụ không hợp lệ. \nVui lòng liên hệ với quản trị viên.",
     "ServiceCreated": "Dịch vụ mẫu {{ name }} đã được tạo thành công.",
@@ -3011,7 +3024,6 @@
     "Copied": "Đã sao chép",
     "Copy": "Sao chép"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "Tạo một phiên",
     "CreateFolder": "Tạo thư mục lưu trữ mới",
