+++ resources/i18n/zh-CN.json
@@ -918,6 +918,9 @@
     "CreatedBy": "创建者",
     "Current": "当前",
     "CurrentRevision": "当前 Revision",
+    "CurrentRevisionAvailableDescription": "加载其配置以作为新 Revision 的起点。",
+    "CurrentRevisionAvailableTitle": "当前 Revision 已在生效中",
+    "CustomMode": "高级模式",
     "DeleteDeployment": "删除部署",
     "Deploy": "部署",
     "DeployConfirm": "确定要部署 Revision #{{revisionNumber}} 吗？当前 Revision 将被替换。",
@@ -953,6 +956,7 @@
     "Image": "镜像",
     "LegacyRouteRedirected": "端点 URL 已迁移。您已被重定向到新的部署页面。",
     "LivenessStatus": "Liveness 状态",
+    "LoadCurrentRevision": "加载当前 Revision",
     "ManageAccessTokens": "管理访问令牌",
     "Model": "模型",
     "ModelDefinitionPath": "模型定义文件路径",
@@ -968,10 +972,15 @@
     "NewRevisionWillBeCreatedConfirm": "保存将创建新的 Revision。是否继续？",
     "NoCurrentRevisionDeployed": "尚未部署任何 Revision — 添加 Revision 以激活此服务。",
     "NoDeployments": "未找到部署。",
+    "NoPresetsAvailable": "没有可用的部署预设",
+    "NoPresetsAvailableDescription": "没有可用于部署的预设。请改为创建新部署。",
+    "NoPresetsAvailableSwitchToCustom": "没有可用的预设。请切换到自定义模式手动添加 Revision。",
     "NumberOfReplicas": "副本数量",
+    "OpenCreateDeploymentModal": "打开创建部署弹窗",
     "OpenToPublic": "公开",
     "Overview": "概览",
     "Owner": "所有者",
+    "PresetMode": "预设模式",
     "Private": "私有",
     "PrivateDeploymentAlertTitle": "私有部署 — 使用访问令牌访问端点。",
     "Project": "项目",
@@ -1777,6 +1786,7 @@
     "ClearErrors": "清除错误并重试",
     "CloneModelFolderConfirm": "要运行模型服务，需要克隆文件夹。\n您想将其克隆为<strong>\"{{folderName}}\"</strong>并启动服务吗？",
     "Controls": "控制装置",
+    "CreateNewDeploymentWithPreset": "使用预设创建新部署",
     "CreatedAt": "创建于",
     "CreatingService": "正在创建服务...",
     "CurrentRevisionTitle": "当前版本",
@@ -1815,8 +1825,12 @@
     "Image": "图片",
     "InitialDelay": "初始延迟",
     "InitialDelayTooltip": "服务启动后，进行第一次健康检查前等待的秒数。",
+    "Interval": "检查间隔 (s)",
+    "IntervalTooltip": "连续健康检查尝试之间的间隔秒数。",
     "MaxRetries": "最大重试次数",
     "MaxRetriesTooltip": "服务被判定为失败之前，健康检查的最大尝试次数。",
+    "MaxWaitTime": "最大等待时间 (s)",
+    "MaxWaitTimeTooltip": "服务响应单次健康检查的最大秒数，超过此时间将视为失败。",
     "ModelAndServingConfiguration": "模型与服务配置",
     "ModelDefinition": "模型定义",
     "ModelDefinitionPath": "模型定义文件路径",
@@ -1903,7 +1917,6 @@
     "RuntimeVariantNotSpecified": "服务定义中未指定 runtime variant。请选择部署预设以继续。",
     "RuntimeVariantSelectionRequired": "有多个 runtime variant 可供选择。请选择部署预设以继续。",
     "SeeContainerLogs": "查看容器日志",
-    "SelectDeploymentPreset": "选择部署预设",
     "ServiceAlreadyExists": "服务名称已经存在",
     "ServiceConfigurationNotValid": "服务配置无效。\n请联系管理员。",
     "ServiceCreated": "模型服务 {{ name }} 已成功创建。",
@@ -3011,7 +3024,6 @@
     "Copied": "已复制",
     "Copy": "复制"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "创建会话",
     "CreateFolder": "创建新的存储文件夹",
