+++ resources/i18n/zh-TW.json
@@ -919,6 +919,9 @@
     "CreatedBy": "建立者",
     "Current": "目前",
     "CurrentRevision": "目前 Revision",
+    "CurrentRevisionAvailableDescription": "載入其設定，以作為新 Revision 的起點。",
+    "CurrentRevisionAvailableTitle": "目前已有 Revision 生效中",
+    "CustomMode": "進階模式",
     "DeleteDeployment": "刪除部署",
     "Deploy": "部署",
     "DeployConfirm": "確定要部署 Revision #{{revisionNumber}} 嗎？當前 Revision 將被替換。",
@@ -954,6 +957,7 @@
     "Image": "映像",
     "LegacyRouteRedirected": "端點 URL 已遷移。您已被重新導向到新的部署頁面。",
     "LivenessStatus": "Liveness 狀態",
+    "LoadCurrentRevision": "載入目前 Revision",
     "ManageAccessTokens": "管理存取權杖",
     "Model": "模型",
     "ModelDefinitionPath": "模型定義檔案路徑",
@@ -969,10 +973,15 @@
     "NewRevisionWillBeCreatedConfirm": "儲存將建立新的 Revision。是否繼續？",
     "NoCurrentRevisionDeployed": "尚未部署任何 Revision — 新增 Revision 以啟動此服務。",
     "NoDeployments": "未找到部署。",
+    "NoPresetsAvailable": "沒有可用的部署預設",
+    "NoPresetsAvailableDescription": "沒有可用於部署的預設。請改為建立新部署。",
+    "NoPresetsAvailableSwitchToCustom": "沒有可用的預設。請切換至自訂模式手動新增 Revision。",
     "NumberOfReplicas": "副本數量",
+    "OpenCreateDeploymentModal": "開啟建立部署視窗",
     "OpenToPublic": "公開",
     "Overview": "概覽",
     "Owner": "擁有者",
+    "PresetMode": "預設模式",
     "Private": "私有",
     "PrivateDeploymentAlertTitle": "私有部署 — 使用存取權杖存取端點。",
     "Project": "專案",
@@ -1778,6 +1787,7 @@
     "ClearErrors": "清除错误并重试",
     "CloneModelFolderConfirm": "要運行模型服務，需要克隆文件夾。您想將其克隆為<strong>\"{{folderName}}\"</strong>並啟動服務嗎？",
     "Controls": "控制装置",
+    "CreateNewDeploymentWithPreset": "使用預設建立新部署",
     "CreatedAt": "创建于",
     "CreatingService": "正在建立服務...",
     "CurrentRevisionTitle": "目前版本",
@@ -1817,8 +1827,12 @@
     "Image": "图片",
     "InitialDelay": "初始延遲",
     "InitialDelayTooltip": "服務啟動後，執行第一次健康檢查前等待的秒數。",
+    "Interval": "檢查間隔 (s)",
+    "IntervalTooltip": "連續健康檢查嘗試之間的間隔秒數。",
     "MaxRetries": "最大重試次數",
     "MaxRetriesTooltip": "服務被判定為失敗之前，健康檢查的最大嘗試次數。",
+    "MaxWaitTime": "最大等待時間 (s)",
+    "MaxWaitTimeTooltip": "服務回應單次健康檢查的最大秒數，超過此時間將視為失敗。",
     "ModelAndServingConfiguration": "模型與服務配置",
     "ModelDefinition": "模型定義",
     "ModelDefinitionPath": "模型定義檔路徑",
@@ -1905,7 +1919,6 @@
     "RuntimeVariantNotSpecified": "服務定義中未指定 runtime variant。請選擇部署預設以繼續。",
     "RuntimeVariantSelectionRequired": "有多個 runtime variant 可供選擇。請選擇部署預設以繼續。",
     "SeeContainerLogs": "查看容器日誌",
-    "SelectDeploymentPreset": "選擇部署預設",
     "ServiceAlreadyExists": "服務名稱已經存在",
     "ServiceConfigurationNotValid": "服務配置無效。\n請聯繫管理員。",
     "ServiceCreated": "模型服務 {{ name }} 已成功建立。",
@@ -3013,7 +3026,6 @@
     "Copied": "已複製",
     "Copy": "複製"
   },
-  "ssession": {},
   "start": {
     "CreateASession": "创建会话",
     "CreateFolder": "建立新的儲存資料夾",
